Split constant folding and partial shape eval out of onnxsim.cpp
onnxsim.cpp had grown to ~4400 lines mixing several loosely-coupled
concerns in one translation unit. Extract two cohesive, self-contained
pieces into their own files, following the existing sym_shape_infer.cpp
/ tensor_pool.cpp split pattern:

- constant_folding.{h,cpp}: op-determinism predicates, the built-in ORT
  executor, RunOps/RunOpsAndAddInitializers, and both the ModelProto- and
  Graph-native constant folders (GetConstantNodes*/FoldGroup*/
  _FoldConstant*), plus Optimize() and the shared Config.
- partial_shape_eval.{h,cpp}: ONNX data-propagation + native symbolic
  shape/value evaluation (_EvalPartialShape*), and
  EliminateZeroRnnInitialState.

onnxsim.cpp keeps the top-level orchestration (Simplify/SimplifyPath,
OptAndShapeOnGraph, quantization entry points, model-hygiene passes) and
shrinks from ~4370 to ~1710 lines. No behavior change; verified by
building the onnxsim library, the ORT-off wheel path (matching
setup.py's actual build), and running onnxsim.simplify() end to end
against models exercising constant folding, the Shape/Gather/Reshape
partial-shape-eval path, and RNN zero-initial-state elimination.
