+++ onnxsim/onnxsim.cpp
@@ -33,6 +33,7 @@
 // core/common/endian.h (which neither ships).
 #include "onnxruntime_cxx_api.h"
 #endif
+#include "constant_folding.h"
 #include "contrib_schemas.h"
 #include "custom_optimizer_passes.h"
 #include "dlpack_bridge.h"
@@ -49,6 +50,7 @@
 #include "onnxoptimizer/optimize.h"
 #include "onnxoptimizer/passes/cse_util.h"
 #include "onnxoptimizer/passes/logging.h"
+#include "partial_shape_eval.h"
 #include "passes/quantize_conv_common.h"
 #include "passes/quantize_fp16.h"
 #include "passes/quantize_matmul_common.h"
@@ -56,2669 +58,6 @@
 #include "profiler.h"
 #include "sym_shape_infer.h"
 #include "sym_value_eval.h"
-
-struct Config {
-  std::vector<std::string> optimizer_passes;
-  // default value is max
-  size_t tensor_size_threshold = -1;
-  // Whether graph initializers are treated as constant tensors. When false,
-  // constant folding does not seed its constant set with initializer names, so
-  // nodes rooted only at initializers are left unfolded (see GetConstantNodes),
-  // and the onnx optimizer is told to leave initializer-backed weights alone
-  // too (see Optimize). Constant nodes are always constant.
-  bool initializers_as_constants = true;
-};
-
-Config config;
-
-bool IsOfficialOp(const std::string& domain, const std::string& op) {
-  if (domain != "ai.onnx" && domain != "ai.onnx.ml" && !domain.empty()) {
-    return false;
-  }
-  // these experimental ops were in onnx default domain but are no
-  // longer supported by onnx now.
-  static std::set<std::string> experimental_ops = {"ATen",
-                                                   "Affine",
-                                                   "ConstantFill",
-                                                   "Crop",
-                                                   "DynamicSlice",
-                                                   "GRUUnit",
-                                                   "GivenTensorFill",
-                                                   "ImageScaler",
-                                                   "ParametricSoftplus",
-                                                   "Scale",
-                                                   "ScaledTanh"};
-  return experimental_ops.find(op) == experimental_ops.end();
-}
-
-// Correct the determinism metadata of operators ONNX mis-annotates, so the
-// ordinary ``IsDeterministic`` check below can fold them.
-//
-// ``OpSchema::GetNodeDeterminism`` infers a *function* op's determinism from
-// the ops in its function body, and reports ``NonDeterministic`` for a body
-// that contains a subgraph-carrying op (``Loop``/``If``/``Scan``) and
-// ``Unknown`` for a context-dependent function -- neither of which means the op
-// is actually random. ``Range`` is the canonical victim: its body is a ``Loop``
-// (opset < 27) or a context-dependent function (opset >= 27), so it is reported
-// non-deterministic even though its output is a pure function of its inputs. It
-// is then never constant-folded, which in turn strands whole static subgraphs
-// built on top of it -- e.g. the ``Range -> Slice -> Reshape -> Expand ->
-// Unsqueeze -> Concat`` attention-mask construction (and the neighbouring
-// ``ScatterND`` chains) in Swin-style models, leaving hundreds of constant
-// nodes that other simplifiers fold away.
-//
-// Rather than second-guess the determinism query in ``IsDeterministic``, fix
-// the source data: mark these genuinely-deterministic ops ``Deterministic`` on
-// their registered schemas (every version in the registry's history). The
-// registry returns pointers into its own storage, so this updates the metadata
-// in place.
-void FixupSchemaDeterminism() {
-  static std::once_flag once;
-  std::call_once(once, [] {
-    // Deterministic default-domain ops whose schema determinism ONNX infers
-    // (incorrectly, for folding purposes) from a function body.
-    static const std::set<std::string> deterministic_ops = {"Range"};
-    for (const auto& schema :
-         onnx::OpSchemaRegistry::get_all_schemas_with_history()) {
-      if (!schema.domain().empty() || !deterministic_ops.count(schema.Name())) {
-        continue;
-      }
-      const onnx::OpSchema* registered = onnx::OpSchemaRegistry::Schema(
-          schema.Name(), schema.since_version(), schema.domain());
-      if (registered != nullptr) {
-        const_cast<onnx::OpSchema*>(registered)
-            ->SetNodeDeterminism(
-                onnx::OpSchema::NodeDeterminism::Deterministic);
-      }
-    }
-  });
-}
-
-bool IsDeterministic(const std::string& domain, const std::string& op,
-                     int opset_version) {
-  // Query the determinism attribute of the operator schema instead of
-  // maintaining a hardcoded list of non-deterministic ops. See
-  // https://github.com/onnx/onnx/pull/7176. Operators ONNX mis-annotates for
-  // constant-folding purposes (e.g. ``Range``) have their metadata corrected by
-  // FixupSchemaDeterminism(), which Simplify() runs before folding.
-  //
-  // The ONNX operator schema registry stores the default ONNX domain as an
-  // empty string.
-  const std::string& lookup_domain = domain == "ai.onnx" ? "" : domain;
-  const auto* schema =
-      onnx::OpSchemaRegistry::Schema(op, opset_version, lookup_domain);
-  if (schema == nullptr) {
-    // Unknown op. Assume it is not deterministic.
-    return false;
-  }
-  // Only fold ops that are known to be deterministic. Ops whose determinism
-  // cannot be statically determined (e.g. context-dependent functions) are
-  // treated as non-deterministic to be safe.
-  return schema->GetNodeDeterminism() ==
-         onnx::OpSchema::NodeDeterminism::Deterministic;
-}
-
-bool IsQDQ(const std::string& domain, const std::string& op) {
-  if (domain == "ai.onnx" || domain.empty()) {
-    return op == "QuantizeLinear" || op == "DequantizeLinear";
-  }
-  return false;
-}
-
-// Returns a reference into `model`'s own initializer list rather than a copy:
-// callers that only read the tensor (the common case) avoid deep-copying its
-// raw_data bytes just to look it up.
-const onnx::TensorProto& FindInitializerByName(const onnx::ModelProto& model,
-                                               const std::string& name) {
-  for (const auto& initializer : model.graph().initializer()) {
-    if (initializer.name() == name) {
-      return initializer;
-    }
-  }
-  throw std::invalid_argument("no initializer " + name);
-}
-
-auto FindValueInfoProtoByName(const onnx::ModelProto& model,
-                              const std::string& name) {
-  for (const auto& vi : model.graph().value_info()) {
-    if (vi.name() == name) {
-      return vi;
-    }
-  }
-  for (const auto& initializer : model.graph().initializer()) {
-    if (initializer.name() == name) {
-      onnx::ValueInfoProto vi;
-      for (const auto& dim : initializer.dims()) {
-        vi.mutable_type()
-            ->mutable_tensor_type()
-            ->mutable_shape()
-            ->add_dim()
-            ->set_dim_value(dim);
-      }
-      vi.mutable_type()->mutable_tensor_type()->set_elem_type(
-          initializer.data_type());
-      vi.set_name(name);
-      return vi;
-    }
-  }
-  throw std::invalid_argument("no value info " + name);
-}
-
-#ifdef ONNXSIM_HAS_ORT
-// The TensorProto<->Ort::Value converters that used to live here have moved to
-// dlpack_bridge.h and now exchange data through DLManagedTensor:
-//   * inputs: onnxsim::dlpack::BorrowAsOrtValue wraps the feed buffer with the
-//     borrowing CreateTensor overload -- no copy in;
-//   * outputs: onnxsim::dlpack::FromOrtValue moves ORT's own output allocation
-//     into the returned tensor -- no copy out (and no per-element add_*_data).
-
-std::shared_ptr<Ort::Env> GetEnv() {
-  static std::shared_ptr<Ort::Env> env = std::make_shared<Ort::Env>();
-  return env;
-}
-
-// Turn on ONNX Runtime's own per-operator session profiler when
-// ONNXSIM_ORT_PROFILE is set, and return whether it was enabled. This is
-// separate from onnxsim's span profiler (ONNXSIM_PROFILE): it makes each
-// constant-folding session dump ONNX Runtime's detailed per-kernel Chrome
-// trace. The variable names a file prefix (ONNX Runtime writes one
-// ``<prefix>_<timestamp>.json`` per session); the truthy shorthands select a
-// default prefix, mirroring ONNXSIM_PROFILE.
-bool EnableOrtProfilingFromEnv(Ort::SessionOptions& sess_opts) {
-  // Merging (ONNXSIM_MERGE_ORT_PROFILE) also needs the per-session traces, and
-  // writes them to an intermediate prefix that Finish() folds in and deletes.
-  const bool merging = onnxsim::Profiler::Instance().merge_ort_traces();
-  const char* env = std::getenv("ONNXSIM_ORT_PROFILE");
-  if (env == nullptr && !merging) {
-    return false;
-  }
-  std::string prefix;
-  if (merging) {
-    prefix = "onnxsim_ort_merge_tmp";
-  } else {
-    prefix = env;
-    if (prefix.empty() || prefix == "1" || prefix == "true" || prefix == "on" ||
-        prefix == "yes") {
-      prefix = "onnxsim_ort_profile";
-    }
-  }
-#ifdef _WIN32
-  // ORTCHAR_T is wchar_t on Windows; widen the (ASCII) prefix for the API.
-  std::wstring wprefix(prefix.begin(), prefix.end());
-  sess_opts.EnableProfiling(wprefix.c_str());
-#else
-  sess_opts.EnableProfiling(prefix.c_str());
-#endif
-  return true;
-}
-
-struct CppModelExecutor : public ModelExecutor {
-  std::vector<DLManagedTensorPtr> Run(
-      const onnx::ModelProto& model,
-      const std::vector<const DLManagedTensor*>& inputs) const override {
-    // The RunOps call site already profiles each fold group's session run as a
-    // single ``OrtSession`` span (see RunOps); for the built-in executor break
-    // that down further into ``OrtSessionInit`` (building the session, where
-    // ONNX Runtime loads the graph and usually the dominant cost) and
-    // ``OrtSessionRun`` (the inference). All ProfiledScopes are no-ops unless
-    // ONNXSIM_PROFILE is set, so this adds nothing otherwise.
-    std::vector<const char*> input_name_ptrs;
-    std::vector<const char*> output_name_ptrs;
-    std::transform(
-        model.graph().input().begin(), model.graph().input().end(),
-        std::back_inserter(input_name_ptrs),
-        [](const onnx::ValueInfoProto& x) { return x.name().c_str(); });
-    std::transform(
-        model.graph().output().begin(), model.graph().output().end(),
-        std::back_inserter(output_name_ptrs),
-        [](const onnx::ValueInfoProto& x) { return x.name().c_str(); });
-    Ort::SessionOptions sess_opts;
-    sess_opts.SetLogSeverityLevel(3);
-    sess_opts.SetGraphOptimizationLevel(ORT_DISABLE_ALL);
-    // This executor exists only to run constant-folding's throwaway
-    // fold-group sub-models (see RunOps above) -- never a full-size
-    // correctness check -- and each session here runs exactly once. So:
-    //  - Memory-pattern planning, which pays off across *repeated* Run()
-    //    calls, buys nothing for a session used once; skip planning it.
-    //  - A fresh intra-op thread pool sized to the machine's CPU count is
-    //    spun up (and joined) on every session construction, which happens
-    //    once per fold group per fixed-point round -- often hundreds of times
-    //    per large model. For the shape/index ops typical of a fold group
-    //    that spin-up/join is pure overhead, and it is where most of a fold
-    //    session's time goes (see the ``OrtSessionInit`` span above).
-    //    Running single-threaded skips it.
-    sess_opts.DisableMemPattern();
-    sess_opts.SetIntraOpNumThreads(1);
-    sess_opts.SetInterOpNumThreads(1);
-    const bool ort_profiling = EnableOrtProfilingFromEnv(sess_opts);
-    std::string model_str = model.SerializeAsString();
-    Ort::Session session{nullptr};
-    {
-      onnxsim::ProfiledScope init_scope("OrtSessionInit");
-      session = Ort::Session(*GetEnv(), model_str.data(), model_str.size(),
-                             sess_opts);
-    }
-    Ort::RunOptions run_opts;
-    run_opts.SetRunLogSeverityLevel(3);
-    // Borrow each feed's buffer directly into an Ort::Value -- no copy. The
-    // DLManagedTensors are owned by the caller (RunOps) and outlive this call,
-    // so the borrowed pointers stay valid through session.Run.
-    std::vector<Ort::Value> input_tensors;
-    input_tensors.reserve(inputs.size());
-    for (const DLManagedTensor* in : inputs) {
-      input_tensors.push_back(onnxsim::dlpack::BorrowAsOrtValue(in->dl_tensor));
-    }
-    std::vector<Ort::Value> output_tensors;
-    {
-      onnxsim::ProfiledScope run_scope("OrtSessionRun");
-      output_tensors =
-          session.Run(run_opts, input_name_ptrs.data(), input_tensors.data(),
-                      input_tensors.size(), output_name_ptrs.data(),
-                      output_name_ptrs.size());
-    }
-    if (ort_profiling) {
-      // Flush ONNX Runtime's profiling trace for this session to disk. When
-      // merging, hand its path to the profiler so Finish() folds it into the
-      // onnxsim trace (and deletes the intermediate file).
-      Ort::AllocatorWithDefaultOptions allocator;
-      Ort::AllocatedStringPtr profile_file =
-          session.EndProfilingAllocated(allocator);
-      if (onnxsim::Profiler::Instance().merge_ort_traces() &&
-          profile_file != nullptr) {
-        onnxsim::Profiler::Instance().AddOrtTracePath(profile_file.get());
-      }
-    }
-
-    // Hand ORT's own output buffers out as DLManagedTensors: FromOrtValue moves
-    // each Ort::Value into the managed tensor, so nothing is copied here (the
-    // one unavoidable copy happens when RunOps bakes the result into the
-    // model's initializers as raw_data).
-    std::vector<DLManagedTensorPtr> outputs;
-    outputs.reserve(output_tensors.size());
-    for (auto& v : output_tensors) {
-      outputs.emplace_back(onnxsim::dlpack::FromOrtValue(std::move(v)));
-    }
-    return outputs;
-  }
-};
-
-std::shared_ptr<const ModelExecutor> GetBuiltinModelExecutor() {
-  static std::shared_ptr<const ModelExecutor> executor =
-      std::make_shared<CppModelExecutor>();
-  return executor;
-}
-
-void InitEnv() { GetEnv(); }
-#else
-void InitEnv() {
-  // do nothing
-}
-#endif
-
-// Fold a group of const nodes together by building one sub-model that produces
-// all of their outputs, running it through `executor` in a single Session, and
-// returning the resulting tensors (named, in group order). Folding many nodes
-// per Session collapses what used to be one Session construction per node into
-// one per group, which is the dominant cost of constant folding. `ops` must be
-// in topological order (as produced by GetConstantNodes); a group may therefore
-// contain nodes that consume the outputs of earlier nodes in the same group --
-// such tensors stay internal to the sub-model instead of becoming feeds.
-//
-// `deferred_producers` maps the output name of every "deferred" node (a
-// ConstantOfShape or Constant->Expand that is logically constant but was not
-// materialized because it would produce a large tensor, see GetConstantNodes)
-// to the node itself. When an input of a grouped node is such a deferred
-// output, the producing node is inlined into the sub-model instead of being
-// looked up as an already-materialized initializer. This is applied
-// transitively, so a whole chain (e.g. ConstantOfShape -> Expand -> Reshape)
-// runs together inside the executor: the large intermediate tensors are
-// computed transiently and only the (smaller) grouped outputs are returned to
-// be stored as initializers.
-std::vector<onnx::TensorProto> RunOps(
-    const ModelExecutor& executor, const onnx::ModelProto& model,
-    const std::vector<const onnx::NodeProto*>& ops,
-    const std::unordered_map<std::string, const onnx::NodeProto*>&
-        deferred_producers) {
-  std::vector<std::string> input_names;
-  // Pointers borrow directly from `model`'s own initializers -- `model` is
-  // const and not touched again until every pointer here has been consumed
-  // (by the DLPack bridge below, within this same call), so this avoids
-  // deep-copying each constant input's raw_data just to feed the executor.
-  std::vector<const onnx::TensorProto*> input_tps;
-  // Names already emitted as a feed or an initializer of the sub-model, so a
-  // constant shared by several grouped nodes is added exactly once.
-  std::set<std::string> seen_inputs;
-
-  // Build the throwaway sub-model on an arena. RunOp is called once per
-  // foldable node -- often thousands of times across a fixed-point run -- and
-  // each call copies initializers and nodes into `op_model`, so the message is
-  // a deep tree of nested sub-messages (NodeProto, TensorProto, ValueInfoProto,
-  // dims, ...). Without an arena, destroying it walks that whole tree freeing
-  // each sub-message individually; on an arena the entire tree is released in
-  // one bulk free when `arena` goes out of scope. `Create` propagates the arena
-  // pointer to every `add_*`/`mutable_*` sub-message -- that propagation is
-  // what makes the teardown cheap. (Older protobuf spelled this
-  // arena-propagating form `Arena::CreateMessage`, but that alias was
-  // deprecated in protobuf 5.x and removed in 6.x -- the floor the bundled ONNX
-  // now requires -- so `Create` is the modern equivalent for message types.)
-  // The sub-model is strictly local: it is never Swap'd or moved into `model`,
-  // and the executor returns its outputs in a separate std::vector that does
-  // not live on this arena, so the arena can be torn down on return without
-  // dangling anything the caller keeps.
-  google::protobuf::Arena arena;
-  onnx::ModelProto& op_model =
-      *google::protobuf::Arena::Create<onnx::ModelProto>(&arena);
-  // Spans the sub-model-construction phase below (building op_model: copying
-  // each grouped node and its constant inputs into the throwaway sub-model)
-  // -- not lexically scoped via ProfiledScope's RAII since input_names/
-  // input_tps/output_names, populated in this phase, are also read by the
-  // DLPack-bridging and output-materialization phases after it ends. See
-  // ONNXSIM_PROFILE's own doc comment for what "the tensor copying inside
-  // constant folding" actually covers.
-  onnxsim::Profiler::Instance().Begin("BuildSubModel");
-  op_model.set_ir_version(model.ir_version());
-  for (const auto& x : model.opset_import()) {
-    *op_model.add_opset_import() = x;
-  }
-
-  // Outputs produced by a node in the group: these are computed inside the
-  // sub-model, so a grouped node consuming one must not treat it as an external
-  // constant feed.
-  std::set<std::string> internal_outputs;
-  for (const auto* op : ops) {
-    for (const auto& output : op->output()) {
-      internal_outputs.insert(output);
-    }
-  }
-
-  // Post-order traversal: emit every deferred producer before its consumer, and
-  // each grouped node in topological order, so the sub-model stays
-  // topologically sorted. Each node is included at most once even when several
-  // consumers share it.
-  std::set<const onnx::NodeProto*> included;
-  std::function<void(const onnx::NodeProto&)> include_node =
-      [&](const onnx::NodeProto& node) {
-        if (!included.insert(&node).second) {
-          return;
-        }
-        for (const auto& input : node.input()) {
-          // skip "" which represents the unset optional input
-          if (input.empty()) {
-            continue;
-          }
-          // Produced by another node in the group: it is an intermediate of the
-          // sub-model, not an external input.
-          if (internal_outputs.find(input) != internal_outputs.end()) {
-            continue;
-          }
-          auto deferred_iter = deferred_producers.find(input);
-          if (deferred_iter != deferred_producers.end()) {
-            // Produced by a deferred node: inline it rather than treating the
-            // (unmaterialized) output as an external constant input.
-            include_node(*deferred_iter->second);
-            continue;
-          }
-          if (!seen_inputs.insert(input).second) {
-            continue;
-          }
-          const auto& in_tp = FindInitializerByName(model, input);
-          if (in_tp.dims().size() == 1 && in_tp.dims()[0] == 0) {
-            *op_model.mutable_graph()->add_initializer() = in_tp;
-            continue;
-          }
-          input_names.push_back(input);
-          input_tps.push_back(&in_tp);
-        }
-        *op_model.mutable_graph()->add_node() = node;
-      };
-  for (const auto* op : ops) {
-    include_node(*op);
-  }
-
-  for (const auto& x : input_names) {
-    // skip "" which represents the unset optional input
-    if (x.empty()) {
-      continue;
-    }
-    *op_model.mutable_graph()->add_input() = FindValueInfoProtoByName(model, x);
-  }
-  // Mark every grouped output as a graph output so the single Run materializes
-  // all of them. `output_names` records them in graph-output order, which is
-  // the order the executor returns the tensors in.
-  std::vector<std::string> output_names;
-  for (const auto* op : ops) {
-    for (const auto& x : op->output()) {
-      onnx::ValueInfoProto vi;
-      // In principle output ValueInfoProto must have type. But it is not
-      // checked.
-      vi.set_name(x);
-      *op_model.mutable_graph()->add_output() = vi;
-      output_names.push_back(x);
-    }
-  }
-  onnxsim::Profiler::Instance().End();  // BuildSubModel
-
-  using namespace ONNX_NAMESPACE::optimization;
-  VLOG(1) << "Running " << ops.size() << " node(s) as one batch";
-  // Constant folding's actual work is running each fold group's sub-model
-  // through the model executor -- an ONNX Runtime session. Profile that run so
-  // it shows up in the trace nested under FoldConstant. This is the one spot
-  // common to every executor (the built-in ONNX Runtime one and the Python
-  // trampoline that Python's simplify() injects), so the session run is
-  // profiled regardless of binding. The ProfiledScope is a no-op unless
-  // ONNXSIM_PROFILE is set.
-  // Bridge to the DLPack executor boundary. Each feed borrows its initializer's
-  // buffer (no copy); `input_tps` is fully built above and not mutated again,
-  // so the borrowed pointers stay valid. `input_dls` owns the managed tensors
-  // and must outlive the executor call (the executor borrows them). Outputs
-  // come back as DLManagedTensors and are baked into TensorProto raw_data here
-  // -- the single, unavoidable copy, since folded results become model
-  // initializers.
-  std::vector<DLManagedTensorPtr> input_dls;
-  std::vector<const DLManagedTensor*> input_ptrs;
-  {
-    onnxsim::ProfiledScope dlpack_input_scope("DLPackInputBridge");
-    input_dls.reserve(input_tps.size());
-    for (const auto* tp : input_tps) {
-      input_dls.emplace_back(onnxsim::dlpack::FromTensorProtoBorrowing(*tp));
-    }
-    input_ptrs.reserve(input_dls.size());
-    for (const auto& p : input_dls) input_ptrs.push_back(p.get());
-  }
-
-  std::vector<DLManagedTensorPtr> output_dls;
-  {
-    onnxsim::ProfiledScope session_scope("OrtSession");
-    output_dls = executor.Run(op_model, input_ptrs);
-  }
-  std::vector<onnx::TensorProto> output_tps;
-  onnxsim::ProfiledScope dlpack_output_scope("DLPackOutputCopy");
-  output_tps.reserve(output_dls.size());
-  for (size_t i = 0; i < output_dls.size(); i++) {
-    output_tps.push_back(onnxsim::dlpack::ToTensorProto(
-        output_dls[i]->dl_tensor,
-        i < output_names.size() ? output_names[i] : std::string()));
-  }
-  return output_tps;
-}
-
-void RunOpsAndAddInitializers(
-    const ModelExecutor& executor, onnx::ModelProto& model,
-    const std::vector<const onnx::NodeProto*>& ops,
-    const std::unordered_map<std::string, const onnx::NodeProto*>&
-        deferred_producers) {
-  const auto output_tps = RunOps(executor, model, ops, deferred_producers);
-  for (const auto& output_tp : output_tps) {
-    *model.mutable_graph()->add_initializer() = output_tp;
-  }
-}
-
-bool HasSubgraph(const onnx::NodeProto& node) {
-  for (const auto& attr : node.attribute()) {
-    if (attr.type() == onnx::AttributeProto::GRAPH ||
-        attr.type() == onnx::AttributeProto::GRAPHS) {
-      return true;
-    }
-  }
-  return false;
-}
-
-size_t size_of_dtype(onnx::TensorProto::DataType dtype) {
-  switch (dtype) {
-    case onnx::TensorProto::DataType::TensorProto_DataType_BOOL:
-    case onnx::TensorProto::DataType::TensorProto_DataType_INT8:
-    case onnx::TensorProto::DataType::TensorProto_DataType_UINT8:
-      return 1;
-    case onnx::TensorProto::DataType::TensorProto_DataType_BFLOAT16:
-    case onnx::TensorProto::DataType::TensorProto_DataType_FLOAT16:
-    case onnx::TensorProto::DataType::TensorProto_DataType_INT16:
-    case onnx::TensorProto::DataType::TensorProto_DataType_UINT16:
-      return 2;
-    case onnx::TensorProto::DataType::TensorProto_DataType_FLOAT:
-    case onnx::TensorProto::DataType::TensorProto_DataType_INT32:
-    case onnx::TensorProto::DataType::TensorProto_DataType_UINT32:
-      return 4;
-    case onnx::TensorProto::DataType::TensorProto_DataType_DOUBLE:
-    case onnx::TensorProto::DataType::TensorProto_DataType_INT64:
-    case onnx::TensorProto::DataType::TensorProto_DataType_UINT64:
-    case onnx::TensorProto::DataType::TensorProto_DataType_COMPLEX64:
-      return 8;
-    case onnx::TensorProto::DataType::TensorProto_DataType_COMPLEX128:
-      return 16;
-    // Don't know the size of string.. Just return 16.
-    case onnx::TensorProto::DataType::TensorProto_DataType_STRING:
-      return 16;
-    default:
-    case onnx::TensorProto::DataType::TensorProto_DataType_UNDEFINED:
-      throw std::invalid_argument("Undefined or unknown datatype");
-  }
-  throw std::invalid_argument("Unknown datatype " + std::to_string(dtype));
-}
-
-bool ProduceLargeTensor(const onnx::ModelProto& model,
-                        const onnx::NodeProto& node, size_t threshold) {
-  std::set<std::string> large_tensor_ops{"Tile", "ConstantOfShape", "Expand"};
-  if (large_tensor_ops.find(node.op_type()) == large_tensor_ops.end()) {
-    return false;
-  }
-  for (const auto& value_info : model.graph().value_info()) {
-    if (value_info.name() == node.output(0)) {
-      size_t size = size_of_dtype(static_cast<onnx::TensorProto::DataType>(
-          value_info.type().tensor_type().elem_type()));
-      for (const auto& dim : value_info.type().tensor_type().shape().dim()) {
-        size *= dim.dim_value();
-      }
-      if (size <= threshold) {
-        return false;
-      }
-    }
-  }
-  // If the output is not in value_info, we assume it is large.
-  // There is a possibility that value_info is presented by the shape inference
-  // later and `ProduceLargeTensor` is called again and returns false at that
-  // time.
-  return true;
-}
-
-// The result of partitioning a graph's nodes for constant folding.
-struct ConstantNodePartition {
-  // Nodes whose output is materialized into an initializer by folding.
-  std::vector<onnx::NodeProto> const_nodes;
-  // Nodes kept in the graph (folded consumers reference their outputs via
-  // initializers, or they are genuinely non-constant runtime nodes).
-  std::vector<onnx::NodeProto> non_const_nodes;
-  // Outputs of "deferred" nodes: ConstantOfShape/Expand nodes whose inputs are
-  // all constant but that were not folded into an initializer because they
-  // would produce a large tensor. They stay in the graph (in non_const_nodes),
-  // yet their outputs are treated as constant so downstream constant nodes stay
-  // foldable; RunOps inlines the producing node into the sub-model it executes,
-  // so the large intermediate tensor is computed transiently and never stored.
-  std::set<std::string> deferred_outputs;
-};
-
-ConstantNodePartition GetConstantNodes(const onnx::ModelProto& model) {
-  // tensor with empty name("") represents the empty value of an optional input
-  // so "" should be treated as a name of a constant tensor.
-  //
-  // A hash set, not a vector: every node's every input is looked up against
-  // this set below (``std::all_of`` over ``const_names``), and the set grows
-  // by every initializer up front plus every folded node's outputs as the
-  // scan proceeds. A vector + linear ``std::find`` makes that lookup
-  // O(constants seen so far) per input, i.e. O(nodes * initializers) overall
-  // on a model with many weights; a hash set makes each lookup O(1) average.
-  std::unordered_set<std::string> const_names{""};
-  ConstantNodePartition partition;
-  auto& const_nodes = partition.const_nodes;
-  auto& non_const_nodes = partition.non_const_nodes;
-  // Seed the constant set with the initializer names, unless the caller asked
-  // for initializers to be treated as non-constant. In that case a node whose
-  // inputs are (only) initializers is not foldable, so its weights are left in
-  // the graph untouched; a node fed by a Constant node still folds because ""
-  // and Constant outputs remain in the constant set.
-  if (config.initializers_as_constants) {
-    for (const auto& x : model.graph().initializer()) {
-      const_names.insert(x.name());
-    }
-  }
-  // Map each domain to its imported opset version so the correct operator
-  // schema can be looked up. The default ONNX domain is normalized to an empty
-  // string, which is how the schema registry stores it.
-  std::unordered_map<std::string, int> domain_to_version;
-  for (const auto& opset : model.opset_import()) {
-    const std::string& domain =
-        opset.domain() == "ai.onnx" ? "" : opset.domain();
-    domain_to_version[domain] = opset.version();
-  }
-  auto opset_version_of = [&domain_to_version](const std::string& domain) {
-    const std::string& key = domain == "ai.onnx" ? "" : domain;
-    auto iter = domain_to_version.find(key);
-    return iter == domain_to_version.end() ? 0 : iter->second;
-  };
-  // node is already topo sorted
-  for (const auto& node : model.graph().node()) {
-    const bool foldable =
-        IsOfficialOp(node.domain(), node.op_type()) &&
-        IsDeterministic(node.domain(), node.op_type(),
-                        opset_version_of(node.domain())) &&
-        !IsQDQ(node.domain(), node.op_type()) && !HasSubgraph(node) &&
-        std::all_of(
-            node.input().begin(), node.input().end(),
-            [&const_names](const auto& x) { return const_names.count(x) > 0; });
-    if (!foldable) {
-      non_const_nodes.push_back(node);
-      continue;
-    }
-    if (!ProduceLargeTensor(model, node, config.tensor_size_threshold)) {
-      // Ordinary constant folding: the output is materialized as an
-      // initializer and the node is dropped.
-      const_names.insert(node.output().begin(), node.output().end());
-      const_nodes.push_back(node);
-      continue;
-    }
-    // Large-tensor op. ConstantOfShape and the foldable Expand (the
-    // "Constant -> Expand" pattern) are folded lazily: the node is kept in the
-    // graph but its output is still treated as constant so consumers keep
-    // folding, and RunOps inlines it into the executor's sub-model at fold time
-    // rather than materializing the large tensor as an initializer. Other
-    // large-tensor ops (e.g. Tile) remain fully excluded from folding.
-    if (node.op_type() == "ConstantOfShape" || node.op_type() == "Expand") {
-      const_names.insert(node.output().begin(), node.output().end());
-      partition.deferred_outputs.insert(node.output().begin(),
-                                        node.output().end());
-    }
-    non_const_nodes.push_back(node);
-  }
-  return partition;
-}
-
-// Recursively collect the names of every tensor consumed as a node input,
-// descending into subgraphs (e.g. the branches of "If" or the body of "Loop").
-// Because ONNX subgraphs can reference tensors from the enclosing scope, an
-// initializer in the main graph may be used only by a node inside a subgraph.
-// Collecting names recursively ensures such initializers are not mistaken for
-// unused ones (issue #174).
-void CollectUsedTensorNames(const onnx::GraphProto& graph,
-                            std::set<std::string>& used) {
-  for (const auto& node : graph.node()) {
-    for (const auto& input : node.input()) {
-      if (!input.empty()) {
-        used.insert(input);
-      }
-    }
-    for (const auto& attr : node.attribute()) {
-      if (attr.has_g()) {
-        CollectUsedTensorNames(attr.g(), used);
-      }
-      for (const auto& subgraph : attr.graphs()) {
-        CollectUsedTensorNames(subgraph, used);
-      }
-    }
-  }
-  // Graph outputs must be kept even if no node consumes them.
-  for (const auto& output : graph.output()) {
-    used.insert(output.name());
-  }
-}
-
-// Remove initializers of the main graph that are no longer referenced by any
-// node (including nodes in subgraphs). Constant folding replaces a subgraph of
-// const ops (e.g. a Transpose on a weight) with a freshly computed initializer,
-// but leaves the original operand initializers in place. Without cleanup those
-// dangling weights are duplicated in the graph, which can push the model past
-// the 2GB protobuf limit before the onnx optimizer gets a chance to remove
-// them (issue #174).
-// Takes `model` by value and mutates it in place rather than copying into a
-// separate `result`: every caller already owns a private, uniquely-held copy
-// by this point and passes it via std::move, so this is a cheap move-in, not
-// a deep copy of the (potentially huge) initializer bytes.
-onnx::ModelProto EliminateUnusedInitializer(onnx::ModelProto model) {
-  std::set<std::string> used;
-  CollectUsedTensorNames(model.graph(), used);
-  // Keep initializers that double as graph inputs (their default value);
-  // dropping them would silently turn them into required inputs.
-  for (const auto& input : model.graph().input()) {
-    used.insert(input.name());
-  }
-
-  google::protobuf::RepeatedPtrField<onnx::TensorProto> kept;
-  for (auto& initializer : *model.mutable_graph()->mutable_initializer()) {
-    if (used.count(initializer.name()) > 0) {
-      *kept.Add() = std::move(initializer);
-    }
-  }
-  model.mutable_graph()->mutable_initializer()->Swap(&kept);
-
-  return model;
-}
-
-// Mutates the model in place; ``onnx::shape_inference::InferShapes`` already
-// works in place, so no extra ModelProto copy is made (the previous ``const&``
-// signature forced a defensive ``CopyFrom`` because the input could not be
-// mutated).
-void _InferShapes(onnx::ModelProto& model) {
-  onnx::shape_inference::InferShapes(model);
-}
-
-// Build a lookup from tensor name to its type, gathering shapes from every
-// place a shape can be declared: value_info (populated by shape inference),
-// graph inputs and graph outputs. Pointers reference `model`, so the map must
-// not outlive it and `model` must not be mutated while the map is in use.
-std::unordered_map<std::string, const onnx::TypeProto*> BuildTypeMap(
-    const onnx::ModelProto& model) {
-  std::unordered_map<std::string, const onnx::TypeProto*> type_map;
-  auto add = [&type_map](const onnx::ValueInfoProto& vi) {
-    if (vi.has_type()) {
-      type_map[vi.name()] = &vi.type();
-    }
-  };
-  for (const auto& vi : model.graph().value_info()) add(vi);
-  for (const auto& vi : model.graph().input()) add(vi);
-  for (const auto& vi : model.graph().output()) add(vi);
-  return type_map;
-}
-
-// Fetch the element type and a fully static shape of `name` from `type_map`.
-// Returns false unless the tensor has a known integer (INT64/INT32) element
-// type and a shape whose every dimension is a fixed value. A rank-0 (scalar)
-// tensor yields an empty `dims` (element count 1).
-bool GetStaticIntTensorInfo(
-    const std::unordered_map<std::string, const onnx::TypeProto*>& type_map,
-    const std::string& name, onnx::TensorProto::DataType& elem_type,
-    std::vector<int64_t>& dims) {
-  auto iter = type_map.find(name);
-  if (iter == type_map.end() || !iter->second->has_tensor_type()) {
-    return false;
-  }
-  const auto& tensor_type = iter->second->tensor_type();
-  elem_type = static_cast<onnx::TensorProto::DataType>(tensor_type.elem_type());
-  if (elem_type != onnx::TensorProto::INT64 &&
-      elem_type != onnx::TensorProto::INT32) {
-    return false;
-  }
-  if (!tensor_type.has_shape()) {
-    // Rank is unknown.
-    return false;
-  }
-  dims.clear();
-  for (const auto& dim : tensor_type.shape().dim()) {
-    if (!dim.has_dim_value()) {
-      return false;
-    }
-    dims.push_back(dim.dim_value());
-  }
-  return true;
-}
-
-// --- Native symbolic shape evaluation (issue #532, milestones M1/M2/M3) ------
-//
-// The ONNX data-propagation path below stalls at any arithmetic over a dynamic
-// dim symbol: it carries a value as a TensorShapeProto whose entries are a
-// concrete int or an *opaque* dim_param string, so a Reshape target like
-// `[batch, 1024, 128]`, or a `Div`/`Where`/`Equal` over the symbol, cannot be
-// evaluated. The dependency-free evaluator in sym_value_eval / sym_shape_infer
-// keeps each dynamic dim as a `SymExpr` and computes the shape algebra. These
-// helpers adapt an `onnx::ModelProto` into the evaluator's plain structs and
-// run M2 (symbolic activation shapes) then M1 (symbolic value evaluation) over
-// it.
-
-// Read one little-endian `T` out of `p`. ONNX defines TensorProto::raw_data as
-// little-endian on every host, so this is a plain byte-wise decode rather than
-// a memcpy into a host integer (which would be correct only on a little-endian
-// machine -- see docs/big-endian.md).
-template <typename T>
-T ReadLittleEndian(const char* p) {
-  std::make_unsigned_t<T> v = 0;
-  for (size_t i = 0; i < sizeof(T); ++i) {
-    v |= static_cast<std::make_unsigned_t<T>>(static_cast<unsigned char>(p[i]))
-         << (8 * i);
-  }
-  return static_cast<T>(v);
-}
-
-// Convert an integer TensorProto (rank 0 or 1, INT64/INT32, inline data) to a
-// SymTensor of concrete values. Returns nullopt for other dtypes/ranks or data
-// kept in an external file.
-std::optional<onnxsim::SymTensor> IntTensorToSymTensor(
-    const onnx::TensorProto& tp) {
-  if (tp.data_location() == onnx::TensorProto::EXTERNAL) return std::nullopt;
-  const auto dt = tp.data_type();
-  if (dt != onnx::TensorProto::INT64 && dt != onnx::TensorProto::INT32)
-    return std::nullopt;
-  if (tp.dims_size() > 1) return std::nullopt;  // rank 0 (scalar) or 1 only
-  const bool scalar = tp.dims_size() == 0;
-  std::vector<int64_t> vals;
-  if (tp.has_raw_data()) {
-    const std::string& raw = tp.raw_data();
-    if (dt == onnx::TensorProto::INT64) {
-      const size_t n = raw.size() / sizeof(int64_t);
-      vals.resize(n);
-      for (size_t i = 0; i < n; ++i)
-        vals[i] = ReadLittleEndian<int64_t>(raw.data() + i * sizeof(int64_t));
-    } else {
-      const size_t n = raw.size() / sizeof(int32_t);
-      vals.resize(n);
-      for (size_t i = 0; i < n; ++i)
-        vals[i] = ReadLittleEndian<int32_t>(raw.data() + i * sizeof(int32_t));
-    }
-  } else if (dt == onnx::TensorProto::INT64) {
-    vals.assign(tp.int64_data().begin(), tp.int64_data().end());
-  } else {
-    vals.assign(tp.int32_data().begin(), tp.int32_data().end());
-  }
-  const int64_t expect = scalar ? 1 : tp.dims(0);
-  if (static_cast<int64_t>(vals.size()) != expect) return std::nullopt;
-  onnxsim::SymTensor t;
-  t.scalar = scalar;
-  for (int64_t v : vals) t.data.emplace_back(v);
-  return t;
-}
-
-// A TypeProto's shape as a SymShape: dim_value -> SymExpr(v), a non-empty
-// dim_param -> its Symbol, an otherwise-unknown dim -> a fresh distinct symbol
-// (so the rank is preserved). Returns nullopt when the rank itself is unknown.
-std::optional<onnxsim::SymShape> TypeProtoToSymShape(
-    const onnx::TypeProto& type, int64_t& fresh) {
-  if (!type.has_tensor_type() || !type.tensor_type().has_shape())
-    return std::nullopt;
-  onnxsim::SymShape shape;
-  for (const auto& dim : type.tensor_type().shape().dim()) {
-    if (dim.has_dim_value())
-      shape.push_back(onnxsim::SymExpr(dim.dim_value()));
-    else if (!dim.dim_param().empty())
-      shape.push_back(onnxsim::SymExpr::Symbol(dim.dim_param()));
-    else
-      shape.push_back(
-          onnxsim::SymExpr::Symbol("seedunk_" + std::to_string(fresh++)));
-  }
-  return shape;
-}
-
-// One node in the evaluator's plain form. A node from a non-default domain gets
-// an empty op_type so no handler matches it (its outputs stay unevaluated).
-onnxsim::SymNode ToSymNode(const onnx::NodeProto& node) {
-  onnxsim::SymNode n;
-  const std::string& domain = node.domain();
-  n.op_type = (domain.empty() || domain == "ai.onnx") ? node.op_type() : "";
-  n.input.assign(node.input().begin(), node.input().end());
-  n.output.assign(node.output().begin(), node.output().end());
-  for (const auto& attr : node.attribute()) {
-    onnxsim::SymAttr a;
-    a.name = attr.name();
-    switch (attr.type()) {
-      case onnx::AttributeProto::INT:
-        a.i = attr.i();
-        break;
-      case onnx::AttributeProto::INTS:
-        a.ints.assign(attr.ints().begin(), attr.ints().end());
-        break;
-      case onnx::AttributeProto::TENSOR:
-        if (auto t = IntTensorToSymTensor(attr.t())) a.t = std::move(*t);
-        break;
-      default:
-        break;
-    }
-    n.attribute.push_back(std::move(a));
-  }
-  return n;
-}
-
-// Run M2 (symbolic activation-shape inference) then M1 (symbolic value
-// evaluation) over `model`, returning every shape-data tensor the evaluator
-// could resolve as a SymTensor (its entries possibly still symbolic).
-std::map<std::string, onnxsim::SymTensor> EvaluateModelSymbolicValues(
-    const onnx::ModelProto& model) {
-  int64_t fresh = 0;
-  std::vector<onnxsim::SymNode> nodes;
-  nodes.reserve(model.graph().node_size());
-  for (const auto& node : model.graph().node())
-    nodes.push_back(ToSymNode(node));
-
-  std::map<std::string, onnxsim::SymTensor> initializers;
-  std::map<std::string, onnxsim::SymShape> shapes_seed;
-  for (const auto& init : model.graph().initializer()) {
-    if (auto t = IntTensorToSymTensor(init)) initializers[init.name()] = *t;
-    onnxsim::SymShape s;  // an initializer's own shape is fully static
-    for (int64_t d : init.dims()) s.emplace_back(d);
-    shapes_seed[init.name()] = std::move(s);
-  }
-  auto seed = [&](const onnx::ValueInfoProto& vi) {
-    if (shapes_seed.count(vi.name()))
-      return;  // keep the concrete initializer shape
-    if (auto s = TypeProtoToSymShape(vi.type(), fresh))
-      shapes_seed[vi.name()] = std::move(*s);
-  };
-  for (const auto& vi : model.graph().input()) seed(vi);
-  for (const auto& vi : model.graph().value_info()) seed(vi);
-  for (const auto& vi : model.graph().output()) seed(vi);
-
-  onnxsim::ShapeGraph sg;
-  sg.node = nodes;
-  sg.value_info = shapes_seed;
-  sg.initializer = initializers;
-
-  onnxsim::SymGraph vg;
-  vg.node = std::move(nodes);
-  vg.initializer = std::move(initializers);
-  vg.shape = onnxsim::InferSymbolicShapes(sg);
-  return onnxsim::EvaluateSymbolicValues(vg);
-}
-
-// --- Graph-native counterparts of the two passes above ----------------------
-//
-// Both InferSymbolicShapes/EvaluateSymbolicValues already operate on the
-// plain, representation-agnostic SymNode/SymTensor/ShapeGraph/SymGraph
-// structs, so only the *builders* need a Graph-native counterpart -- the
-// evaluator logic itself is untouched.
-
-// Graph-native counterpart of IntTensorToSymTensor: identical rank-0/1,
-// INT32/INT64-only rules, reading directly from ir.h's Tensor instead of
-// TensorProto.
-std::optional<onnxsim::SymTensor> IntTensorToSymTensor(const onnx::Tensor& t) {
-  if (t.data_location() == onnx::TensorProto::EXTERNAL) return std::nullopt;
-  const auto dt = t.elem_type();
-  if (dt != onnx::TensorProto::INT64 && dt != onnx::TensorProto::INT32)
-    return std::nullopt;
-  if (t.sizes().size() > 1) return std::nullopt;
-  const bool scalar = t.sizes().empty();
-  std::vector<int64_t> vals;
-  if (t.is_raw_data()) {
-    const std::string& raw = t.raw();
-    if (dt == onnx::TensorProto::INT64) {
-      const size_t n = raw.size() / sizeof(int64_t);
-      vals.resize(n);
-      for (size_t i = 0; i < n; ++i)
-        vals[i] = ReadLittleEndian<int64_t>(raw.data() + i * sizeof(int64_t));
-    } else {
-      const size_t n = raw.size() / sizeof(int32_t);
-      vals.resize(n);
-      for (size_t i = 0; i < n; ++i)
-        vals[i] = ReadLittleEndian<int32_t>(raw.data() + i * sizeof(int32_t));
-    }
-  } else if (dt == onnx::TensorProto::INT64) {
-    vals.assign(t.int64s().begin(), t.int64s().end());
-  } else {
-    vals.assign(t.int32s().begin(), t.int32s().end());
-  }
-  const int64_t expect = scalar ? 1 : t.sizes()[0];
-  if (static_cast<int64_t>(vals.size()) != expect) return std::nullopt;
-  onnxsim::SymTensor r;
-  r.scalar = scalar;
-  for (int64_t v : vals) r.data.emplace_back(v);
-  return r;
-}
-
-// Graph-native counterpart of TypeProtoToSymShape, reading a Value's own
-// elemType()/sizes() -- no ModelProto value_info map needed, since a Value
-// already carries its own current shape/type in the IR.
-std::optional<onnxsim::SymShape> ValueToSymShape(onnx::Value* v,
-                                                 int64_t& fresh) {
-  if (!v->has_sizes()) return std::nullopt;
-  onnxsim::SymShape shape;
-  for (const auto& d : v->sizes()) {
-    if (d.is_int) {
-      shape.push_back(onnxsim::SymExpr(d.dim));
-    } else if (!d.is_unknown) {
-      shape.push_back(onnxsim::SymExpr::Symbol(d.param));
-    } else {
-      shape.push_back(
-          onnxsim::SymExpr::Symbol("seedunk_" + std::to_string(fresh++)));
-    }
-  }
-  return shape;
-}
-
-// Graph-native counterpart of ToSymNode.
-onnxsim::SymNode ToSymNode(onnx::Node* node) {
-  onnxsim::SymNode n;
-  const std::string domain =
-      node->has_domain() ? node->domain() : std::string();
-  n.op_type =
-      (domain.empty() || domain == "ai.onnx") ? node->kind().toString() : "";
-  for (onnx::Value* v : node->inputs()) {
-    n.input.push_back(v->node()->kind() == onnx::kUndefined ? ""
-                                                            : v->uniqueName());
-  }
-  for (onnx::Value* v : node->outputs()) {
-    n.output.push_back(v->uniqueName());
-  }
-  for (onnx::Symbol attr_name : node->attributeNames()) {
-    onnxsim::SymAttr a;
-    a.name = attr_name.toString();
-    switch (node->kindOf(attr_name)) {
-      case onnx::AttributeKind::i:
-        a.i = node->i(attr_name);
-        break;
-      case onnx::AttributeKind::is:
-        a.ints = node->is(attr_name);
-        break;
-      case onnx::AttributeKind::t:
-        if (auto t = IntTensorToSymTensor(node->t(attr_name)))
-          a.t = std::move(*t);
-        break;
-      default:
-        break;
-    }
-    n.attribute.push_back(std::move(a));
-  }
-  return n;
-}
-
-// Graph-native counterpart of EvaluateModelSymbolicValues.
-std::map<std::string, onnxsim::SymTensor> EvaluateGraphSymbolicValues(
-    onnx::Graph& g, const std::vector<onnx::Node*>& node_ptrs) {
-  int64_t fresh = 0;
-  std::vector<onnxsim::SymNode> nodes;
-  nodes.reserve(node_ptrs.size());
-  for (onnx::Node* node : node_ptrs) {
-    if (node->kind() == onnx::kUndefined || node->kind() == onnx::kCaptured)
-      continue;
-    nodes.push_back(ToSymNode(node));
-  }
-
-  std::map<std::string, onnxsim::SymTensor> initializers;
-  std::map<std::string, onnxsim::SymShape> shapes_seed;
-  const auto& inits = g.initializers();
-  const auto& init_names = g.initializer_names();
-  for (size_t i = 0; i < inits.size(); ++i) {
-    if (auto t = IntTensorToSymTensor(*inits[i]))
-      initializers[init_names[i]] = *t;
-    onnxsim::SymShape s;
-    for (int64_t d : inits[i]->sizes()) s.emplace_back(d);
-    shapes_seed[init_names[i]] = std::move(s);
-  }
-  auto seed = [&](onnx::Value* v) {
-    if (shapes_seed.count(v->uniqueName()))
-      return;  // keep the concrete initializer shape
-    if (auto s = ValueToSymShape(v, fresh))
-      shapes_seed[v->uniqueName()] = std::move(*s);
-  };
-  for (onnx::Value* v : g.inputs()) seed(v);
-  for (onnx::Node* node : node_ptrs) {
-    for (onnx::Value* v : node->outputs()) seed(v);
-  }
-  for (onnx::Value* v : g.outputs()) seed(v);
-
-  onnxsim::ShapeGraph sg;
-  sg.node = nodes;
-  sg.value_info = shapes_seed;
-  sg.initializer = initializers;
-
-  onnxsim::SymGraph vg;
-  vg.node = std::move(nodes);
-  vg.initializer = std::move(initializers);
-  vg.shape = onnxsim::InferSymbolicShapes(sg);
-  return onnxsim::EvaluateSymbolicValues(vg);
-}
-
-// Partial shape evaluation (issue #139) via ONNX data propagation.
-//
-// The plain constant folder only folds a node when *all* of its inputs are
-// constant, so shape-computing ops like `Shape` are never folded: their input
-// is an activation. Yet those ops depend solely on shapes, which shape
-// inference knows -- fully or partially -- even when some dimensions stay
-// dynamic.
-//
-// ONNX shape inference can *propagate* those partially known values: with data
-// propagation enabled it fills a DataValueMap mapping each tensor to a
-// TensorShapeProto whose entries are either a concrete dim_value or a symbolic
-// dim_param. Ops across the shape family (Shape, Gather, Slice, Concat,
-// Squeeze/Unsqueeze, Cast, Add/Sub/Mul, ...) participate, so a chain like
-//   Shape([batch, C, H, W]) -> Gather([1, 2, 3])  ==>  [C, H, W]
-// is propagated end to end and comes out fully concrete even though the batch
-// dimension stays dynamic (the mask-rcnn pattern from issue #139).
-//
-// This pass rewrites every node whose lone output has a fully concrete
-// propagated value into a `Constant` node. Downstream ops then fold through the
-// ordinary constant folder, and now-dead nodes are removed by the optimizer.
-// A graph's ``value_info``/``output`` annotations, snapshotted recursively:
-// one entry for the graph itself, followed by one entry per nested subgraph
-// (the branches of an ``If``, the body of a ``Loop``/``Scan``, ...) in the
-// same pre-order that ``RestoreGraphAnnotations`` below walks them back in.
-struct GraphAnnotationSnapshot {
-  google::protobuf::RepeatedPtrField<onnx::ValueInfoProto> value_info;
-  google::protobuf::RepeatedPtrField<onnx::ValueInfoProto> output;
-};
-
-void SnapshotGraphAnnotations(const onnx::GraphProto& graph,
-                              std::vector<GraphAnnotationSnapshot>& out) {
-  out.push_back({graph.value_info(), graph.output()});
-  for (const auto& node : graph.node()) {
-    for (const auto& attr : node.attribute()) {
-      if (attr.has_g()) {
-        SnapshotGraphAnnotations(attr.g(), out);
-      }
-      for (const auto& subgraph : attr.graphs()) {
-        SnapshotGraphAnnotations(subgraph, out);
-      }
-    }
-  }
-}
-
-// ``restore_self`` gates whether `graph` itself (as opposed to the nested
-// subgraphs found in its nodes) gets its snapshot re-applied: callers that
-// intentionally keep a graph's own data-propagated annotations (because
-// something in that graph did get folded using them) still want every nested
-// ``If``/``Loop`` subgraph restored unconditionally -- this pass never folds
-// a subgraph-local node, so a subgraph's annotations are never allowed to
-// stick regardless of what happened at the level above it.
-void RestoreGraphAnnotations(onnx::GraphProto* graph,
-                             const std::vector<GraphAnnotationSnapshot>& snaps,
-                             size_t& idx, bool restore_self = true) {
-  if (restore_self) {
-    *graph->mutable_value_info() = snaps[idx].value_info;
-    *graph->mutable_output() = snaps[idx].output;
-  }
-  idx++;
-  for (auto& node : *graph->mutable_node()) {
-    for (auto& attr : *node.mutable_attribute()) {
-      if (attr.has_g()) {
-        RestoreGraphAnnotations(attr.mutable_g(), snaps, idx);
-      }
-      for (auto& subgraph : *attr.mutable_graphs()) {
-        RestoreGraphAnnotations(&subgraph, snaps, idx);
-      }
-    }
-  }
-}
-
-void _EvalPartialShape(onnx::ModelProto& model) {
-  // This pass runs shape inference with *data propagation* (lenient options)
-  // purely to discover foldable shape values; it must not otherwise change the
-  // model. InferShapes mutates value_info and output types in place --
-  // including inside nested subgraphs (the branches of an ``If``, a ``Loop``
-  // body, ...), since data-propagating shape inference recurses into them
-  // too -- so snapshot those annotations *recursively* and restore them on
-  // the paths that fold nothing, leaving the model byte-for-byte unchanged
-  // (the old code returned the untouched input there). Restoring only the
-  // top-level graph's annotations left a subgraph's lenient, data-propagated
-  // value_info in place even when nothing here was foldable; two sibling
-  // branches of the same ``If`` can independently reuse the same local
-  // output name (e.g. both a decoder's no-past and with-past branches
-  // producing their own "ConstantOfShape_2"), so leaking one branch's
-  // data-propagation shape onto the model let onnxruntime's later merge of
-  // the ``If`` node's branch output types find the two incompatible and
-  // reject the model outright. The snapshot is metadata only -- no tensor
-  // weights -- so it is cheap, unlike the full-model ``CopyFrom`` it
-  // replaces. Restoring also keeps this pass's data-propagation value_info
-  // out of the model, which matters: it differs from the regular
-  // shape-inference pass's value_info, and leaving it behind could make the
-  // outer fixed point oscillate.
-  std::vector<GraphAnnotationSnapshot> saved_annotations;
-  SnapshotGraphAnnotations(model.graph(), saved_annotations);
-  auto restore = [&]() {
-    size_t idx = 0;
-    RestoreGraphAnnotations(model.mutable_graph(), saved_annotations, idx);
-  };
-  // Unlike `restore`, this leaves the top-level graph's own (possibly folded)
-  // annotations alone and only reverts every nested subgraph -- for the path
-  // below where something *was* foldable at the top level, so the top-level
-  // graph's data-propagated value_info is allowed to stick (nothing here
-  // ever folds a subgraph-local node, so a subgraph's annotations must
-  // revert either way).
-  auto restore_subgraphs_only = [&]() {
-    size_t idx = 0;
-    RestoreGraphAnnotations(model.mutable_graph(), saved_annotations, idx,
-                            /*restore_self=*/false);
-  };
-
-  onnx::shape_inference::DataValueMap data_map;
-  try {
-    const onnx::ShapeInferenceOptions options(/*check_type=*/false,
-                                              /*error_mode=*/0,
-                                              /*enable_data_propagation=*/true);
-    onnx::shape_inference::InferShapes(
-        model, onnx::OpSchemaRegistry::Instance(), options, &data_map);
-  } catch (const std::exception&) {
-    // If shape inference fails we simply have no propagated values to exploit.
-    restore();
-    return;
-  }
-
-  // An empty data_map is not a dead end anymore: the native symbolic evaluator
-  // (issue #532) runs further below and can resolve chains ONNX data
-  // propagation could not, so fall through instead of returning. The loops over
-  // data_map simply add nothing, and the final `folded_values && reshape_fixes`
-  // empty check restores the model if the symbolic pass also finds nothing.
-
-  const auto type_map = BuildTypeMap(model);
-
-  // Maps the output of a foldable node to the constant tensor it produces. Each
-  // such node is rewritten into a `Constant` node holding this value.
-  std::unordered_map<std::string, onnx::TensorProto> folded_values;
-
-  for (const auto& node : model.graph().node()) {
-    // Shape-family ops are single-output; only replace a node when its lone
-    // output is fully known, so dropping it can never orphan a second output.
-    if (node.output_size() != 1) {
-      continue;
-    }
-    const std::string& output = node.output(0);
-    auto data_iter = data_map.find(output);
-    if (data_iter == data_map.end()) {
-      continue;
-    }
-
-    // Every element must be statically known. Data propagation represents an
-    // unknown element as a dimension with neither dim_value nor dim_param, so
-    // requiring dim_value on every entry both proves the value is concrete and
-    // filters out activations whose rank alone is known.
-    const onnx::TensorShapeProto& value = data_iter->second;
-    bool fully_known = true;
-    std::vector<int64_t> values;
-    for (const auto& dim : value.dim()) {
-      if (!dim.has_dim_value()) {
-        fully_known = false;
-        break;
-      }
-      values.push_back(dim.dim_value());
-    }
-    if (!fully_known) {
-      continue;
-    }
-
-    // Build the constant tensor with the output's real dtype and shape. The
-    // propagated data is a flat sequence, so require a fully static shape whose
-    // element count matches what was propagated.
-    onnx::TensorProto::DataType elem_type;
-    std::vector<int64_t> dims;
-    if (!GetStaticIntTensorInfo(type_map, output, elem_type, dims)) {
-      continue;
-    }
-    int64_t element_count = 1;
-    for (int64_t d : dims) {
-      element_count *= d;
-    }
-    if (element_count != static_cast<int64_t>(values.size())) {
-      continue;
-    }
-
-    onnx::TensorProto tp;
-    tp.set_data_type(elem_type);
-    for (int64_t d : dims) {
-      tp.add_dims(d);
-    }
-    if (elem_type == onnx::TensorProto::INT64) {
-      for (int64_t v : values) {
-        tp.add_int64_data(v);
-      }
-    } else {
-      for (int64_t v : values) {
-        tp.add_int32_data(static_cast<int32_t>(v));
-      }
-    }
-    folded_values.emplace(output, std::move(tp));
-  }
-
-  // Data propagation for `Reshape` (single dynamic dim). The fully-known folder
-  // above only rewrites a node whose propagated value is entirely concrete, so
-  // a shape tensor that keeps one symbolic entry -- e.g. `[batch, 1024, 128]`
-  // on a graph with a dynamic batch, or `[?, 768]` with a dynamic sequence
-  // length -- is left alone, and with it the whole `Shape -> Gather -> Concat`
-  // subgraph that computes it at runtime. Those single-dynamic-dim reshapes
-  // dominate transformer and speech graphs.
-  //
-  // When a Reshape's shape input propagates to a value with exactly one unknown
-  // entry and all other entries positive constants, materialize the shape as a
-  // constant with the unknown slot set to -1. ONNX Reshape infers the -1 dim
-  // from the total element count, so for every input the result is identical to
-  // the runtime-computed shape, while the shape-producing subgraph becomes dead
-  // and is removed by the optimizer. (Correctness is still gated by onnxsim's
-  // own equivalence check.)
-  struct ReshapeShapeFix {
-    std::string shape_name;
-    onnx::TensorProto shape_tensor;
-  };
-  std::unordered_map<std::string, ReshapeShapeFix> reshape_fixes;
-  for (const auto& node : model.graph().node()) {
-    if (node.op_type() != "Reshape" || node.input_size() < 2 ||
-        node.output_size() != 1) {
-      continue;
-    }
-    auto data_iter = data_map.find(node.input(1));
-    if (data_iter == data_map.end()) {
-      continue;
-    }
-    const onnx::TensorShapeProto& shape_value = data_iter->second;
-    if (shape_value.dim_size() == 0) {
-      continue;
-    }
-    int unknown = 0;
-    bool usable = true;
-    std::vector<int64_t> shape;
-    shape.reserve(shape_value.dim_size());
-    for (const auto& dim : shape_value.dim()) {
-      if (dim.has_dim_value()) {
-        // A non-positive entry is a literal 0 (copy-dim) or an already
-        // materialized -1; leave those for the ordinary folder.
-        if (dim.dim_value() <= 0) {
-          usable = false;
-          break;
-        }
-        shape.push_back(dim.dim_value());
-      } else {
-        ++unknown;
-        shape.push_back(-1);
-      }
-    }
-    // Exactly one unknown dim maps to Reshape's single -1 sentinel. Zero
-    // unknowns is handled by the fully-known folder above; two or more cannot
-    // be expressed with a single -1.
-    if (!usable || unknown != 1) {
-      continue;
-    }
-    onnx::TensorProto tp;
-    tp.set_data_type(onnx::TensorProto::INT64);
-    tp.add_dims(static_cast<int64_t>(shape.size()));
-    for (int64_t v : shape) {
-      tp.add_int64_data(v);
-    }
-    reshape_fixes.emplace(
-        node.output(0),
-        ReshapeShapeFix{node.output(0) + "_dp_shape", std::move(tp)});
-  }
-
-  // Native symbolic evaluation (issue #532). ONNX data propagation above stops
-  // wherever the shape algebra crosses a dynamic-dim symbol; the SymExpr
-  // evaluator resolves those chains. Merge whatever it finds that the ONNX path
-  // did not already cover into the same two rewrite maps, so the shared rewrite
-  // loop below handles both uniformly. Correctness stays gated by check_n.
-  {
-    const auto sym_values = EvaluateModelSymbolicValues(model);
-    // Fully concrete symbolic values fold to a `Constant`, exactly like the
-    // ONNX fully-known folder above (same dtype/shape/element-count checks).
-    for (const auto& node : model.graph().node()) {
-      if (node.output_size() != 1) continue;
-      const std::string& output = node.output(0);
-      if (folded_values.count(output) || reshape_fixes.count(output)) continue;
-      auto sym_iter = sym_values.find(output);
-      if (sym_iter == sym_values.end()) continue;
-      std::vector<int64_t> flat;
-      bool all_concrete = true;
-      for (const auto& e : sym_iter->second.data) {
-        if (e.is_symbolic()) {
-          all_concrete = false;
-          break;
-        }
-        flat.push_back(e.to_int());
-      }
-      if (!all_concrete) continue;
-      onnx::TensorProto::DataType elem_type;
-      std::vector<int64_t> dims;
-      if (!GetStaticIntTensorInfo(type_map, output, elem_type, dims)) continue;
-      int64_t element_count = 1;
-      for (int64_t d : dims) element_count *= d;
-      if (element_count != static_cast<int64_t>(flat.size())) continue;
-      onnx::TensorProto tp;
-      tp.set_data_type(elem_type);
-      for (int64_t d : dims) tp.add_dims(d);
-      if (elem_type == onnx::TensorProto::INT64) {
-        for (int64_t v : flat) tp.add_int64_data(v);
-      } else {
-        for (int64_t v : flat) tp.add_int32_data(static_cast<int32_t>(v));
-      }
-      folded_values.emplace(output, std::move(tp));
-    }
-    // A Reshape whose target has exactly one symbolic entry (plus positive
-    // constants) becomes `[-1, ...]` -- the same rewrite as the ONNX data-prop
-    // Reshape path above, but reached through SymExpr arithmetic.
-    for (const auto& node : model.graph().node()) {
-      if (node.op_type() != "Reshape" || node.input_size() < 2 ||
-          node.output_size() != 1) {
-        continue;
-      }
-      if (reshape_fixes.count(node.output(0))) continue;
-      auto sym_iter = sym_values.find(node.input(1));
-      if (sym_iter == sym_values.end()) continue;
-      const onnxsim::SymTensor& shape_value = sym_iter->second;
-      if (shape_value.scalar || shape_value.data.empty()) continue;
-      int unknown = 0;
-      bool usable = true;
-      std::vector<int64_t> shape;
-      shape.reserve(shape_value.data.size());
-      for (const auto& e : shape_value.data) {
-        if (e.is_symbolic()) {
-          ++unknown;
-          shape.push_back(-1);
-        } else {
-          const int64_t v = e.to_int();
-          if (v <=
-              0) {  // a literal 0 (copy) or -1 is left for the ordinary folder
-            usable = false;
-            break;
-          }
-          shape.push_back(v);
-        }
-      }
-      if (!usable || unknown != 1) continue;
-      onnx::TensorProto tp;
-      tp.set_data_type(onnx::TensorProto::INT64);
-      tp.add_dims(static_cast<int64_t>(shape.size()));
-      for (int64_t v : shape) tp.add_int64_data(v);
-      reshape_fixes.emplace(
-          node.output(0),
-          ReshapeShapeFix{node.output(0) + "_sym_shape", std::move(tp)});
-    }
-  }
-
-  if (folded_values.empty() && reshape_fixes.empty()) {
-    restore();
-    return;
-  }
-  restore_subgraphs_only();
-
-  // Rewrite each foldable node into a `Constant` node in the same position,
-  // keeping the graph topologically sorted. Emitting a `Constant` node (rather
-  // than injecting an initializer) leaves the value in producer form, so the
-  // ordinary constant folder and optimizer decide how to materialize it.
-  google::protobuf::RepeatedPtrField<onnx::NodeProto> original_nodes;
-  original_nodes.Swap(model.mutable_graph()->mutable_node());
-  for (auto& node : original_nodes) {
-    auto iter = node.output_size() == 1 ? folded_values.find(node.output(0))
-                                        : folded_values.end();
-    if (iter != folded_values.end()) {
-      onnx::NodeProto* constant = model.mutable_graph()->add_node();
-      constant->set_name(node.name());
-      constant->set_op_type("Constant");
-      constant->add_output(iter->first);
-      onnx::AttributeProto* attr = constant->add_attribute();
-      attr->set_name("value");
-      attr->set_type(onnx::AttributeProto::TENSOR);
-      *attr->mutable_t() = std::move(iter->second);
-      continue;
-    }
-    auto fix_iter = node.output_size() == 1 ? reshape_fixes.find(node.output(0))
-                                            : reshape_fixes.end();
-    if (fix_iter != reshape_fixes.end()) {
-      // Emit the materialized shape as a Constant just before the Reshape
-      // (preserving topological order), then repoint the Reshape's shape input
-      // at it. The original shape-producing subgraph is now unused and is
-      // cleaned up by the optimizer's dead-node elimination.
-      onnx::NodeProto* shape_const = model.mutable_graph()->add_node();
-      shape_const->set_op_type("Constant");
-      shape_const->add_output(fix_iter->second.shape_name);
-      onnx::AttributeProto* attr = shape_const->add_attribute();
-      attr->set_name("value");
-      attr->set_type(onnx::AttributeProto::TENSOR);
-      *attr->mutable_t() = std::move(fix_iter->second.shape_tensor);
-
-      onnx::NodeProto* reshape = model.mutable_graph()->add_node();
-      *reshape = std::move(node);
-      reshape->set_input(1, fix_iter->second.shape_name);
-      continue;
-    }
-    *model.mutable_graph()->add_node() = std::move(node);
-  }
-}
-
-// Graph-native counterpart of _EvalPartialShape: same two rewrites (fold a
-// fully-known shape-family output to a Constant; materialize a Reshape's
-// shape input as a Constant with a single -1 slot when everything else is
-// concrete), reached without any ModelProto <-> Graph round trip. Built on
-// the extended InferShapesOnGraph (generated_shape_data out-param, see
-// onnx/common/graph_shape_inference.h) instead of onnx's protobuf-based
-// InferShapes. Returns whether anything changed, so the fully Graph-native
-// outer pipeline can use it as a GraphFnChanged step.
-bool _EvalPartialShapeOnGraph(onnx::Graph& g) {
-  // Mirrors _EvalPartialShape's own snapshot/restore of value_info/output:
-  // this pass's own data-propagation inference call must not leave its
-  // (lenient, check_type=false) shape/type conclusions on the graph --
-  // only the two rewrites below should persist. Restore before the rewrite
-  // loop, from a snapshot taken before the inference call.
-  struct ValueTypeSnapshot {
-    int32_t elem_type;
-    bool has_sizes;
-    std::vector<onnx::Dimension> sizes;
-  };
-  std::vector<onnx::Node*> node_ptrs(g.nodes().begin(), g.nodes().end());
-  std::unordered_map<onnx::Value*, ValueTypeSnapshot> snapshot;
-  auto save = [&](onnx::Value* v) {
-    snapshot[v] = {
-        v->elemType(), v->has_sizes(),
-        v->has_sizes() ? v->sizes() : std::vector<onnx::Dimension>{}};
-  };
-  for (onnx::Value* v : g.inputs()) save(v);
-  for (onnx::Node* node : node_ptrs) {
-    for (onnx::Value* v : node->outputs()) save(v);
-  }
-  auto restore = [&]() {
-    for (const auto& [v, s] : snapshot) {
-      v->setElemType(s.elem_type);
-      if (s.has_sizes) {
-        v->setSizes(s.sizes);
-      } else {
-        v->wipeSizes();
-      }
-    }
-  };
-
-  onnx::shape_inference::DataValueMap data_map;
-  {
-    const onnx::ShapeInferenceOptions options(/*check_type=*/false,
-                                              /*error_mode=*/0,
-                                              /*enable_data_propagation=*/true);
-    try {
-      onnx::InferShapesOnGraph(g, options, &data_map);
-    } catch (const std::exception&) {
-      restore();
-      return false;
-    }
-  }
-
-  // Maps the output of a foldable node to the constant tensor it produces.
-  std::unordered_map<std::string, onnx::Tensor> folded_values;
-  for (onnx::Node* node : node_ptrs) {
-    if (node->outputs().size() != 1) continue;
-    onnx::Value* out = node->outputs()[0];
-    auto data_iter = data_map.find(out->uniqueName());
-    if (data_iter == data_map.end()) continue;
-
-    const onnx::TensorShapeProto& value = data_iter->second;
-    bool fully_known = true;
-    std::vector<int64_t> values;
-    for (const auto& dim : value.dim()) {
-      if (!dim.has_dim_value()) {
-        fully_known = false;
-        break;
-      }
-      values.push_back(dim.dim_value());
-    }
-    if (!fully_known) continue;
-
-    const auto elem_type =
-        static_cast<onnx::TensorProto::DataType>(out->elemType());
-    if (elem_type != onnx::TensorProto::INT64 &&
-        elem_type != onnx::TensorProto::INT32) {
-      continue;
-    }
-    if (!out->has_sizes()) continue;
-    std::vector<int64_t> dims;
-    bool dims_known = true;
-    for (const auto& d : out->sizes()) {
-      if (!d.is_int) {
-        dims_known = false;
-        break;
-      }
-      dims.push_back(d.dim);
-    }
-    if (!dims_known) continue;
-    int64_t element_count = 1;
-    for (int64_t d : dims) element_count *= d;
-    if (element_count != static_cast<int64_t>(values.size())) continue;
-
-    onnx::Tensor t;
-    t.elem_type() = elem_type;
-    t.sizes() = dims;
-    if (elem_type == onnx::TensorProto::INT64) {
-      for (int64_t v : values) t.int64s().push_back(v);
-    } else {
-      for (int64_t v : values) t.int32s().push_back(static_cast<int32_t>(v));
-    }
-    folded_values.emplace(out->uniqueName(), std::move(t));
-  }
-
-  // Data propagation for `Reshape` (single dynamic dim) -- see
-  // _EvalPartialShape's own comment for why this matters.
-  struct ReshapeShapeFix {
-    onnx::Tensor shape_tensor;
-  };
-  std::unordered_map<std::string, ReshapeShapeFix> reshape_fixes;
-  for (onnx::Node* node : node_ptrs) {
-    if (node->kind() != onnx::kReshape || node->inputs().size() < 2 ||
-        node->outputs().size() != 1) {
-      continue;
-    }
-    if (folded_values.count(node->outputs()[0]->uniqueName())) continue;
-    auto data_iter = data_map.find(node->inputs()[1]->uniqueName());
-    if (data_iter == data_map.end()) continue;
-    const onnx::TensorShapeProto& shape_value = data_iter->second;
-    if (shape_value.dim_size() == 0) continue;
-    int unknown = 0;
-    bool usable = true;
-    std::vector<int64_t> shape;
-    shape.reserve(shape_value.dim_size());
-    for (const auto& dim : shape_value.dim()) {
-      if (dim.has_dim_value()) {
-        if (dim.dim_value() <= 0) {
-          usable = false;
-          break;
-        }
-        shape.push_back(dim.dim_value());
-      } else {
-        ++unknown;
-        shape.push_back(-1);
-      }
-    }
-    if (!usable || unknown != 1) continue;
-    onnx::Tensor t;
-    t.elem_type() = onnx::TensorProto::INT64;
-    t.sizes() = {static_cast<int64_t>(shape.size())};
-    for (int64_t v : shape) t.int64s().push_back(v);
-    reshape_fixes.emplace(node->outputs()[0]->uniqueName(),
-                          ReshapeShapeFix{std::move(t)});
-  }
-
-  // Native symbolic evaluation (issue #532), merged into the same two
-  // rewrite maps exactly as _EvalPartialShape does.
-  {
-    const auto sym_values = EvaluateGraphSymbolicValues(g, node_ptrs);
-    for (onnx::Node* node : node_ptrs) {
-      if (node->outputs().size() != 1) continue;
-      onnx::Value* out = node->outputs()[0];
-      const std::string& output = out->uniqueName();
-      if (folded_values.count(output) || reshape_fixes.count(output)) continue;
-      auto sym_iter = sym_values.find(output);
-      if (sym_iter == sym_values.end()) continue;
-      std::vector<int64_t> flat;
-      bool all_concrete = true;
-      for (const auto& e : sym_iter->second.data) {
-        if (e.is_symbolic()) {
-          all_concrete = false;
-          break;
-        }
-        flat.push_back(e.to_int());
-      }
-      if (!all_concrete) continue;
-      const auto elem_type =
-          static_cast<onnx::TensorProto::DataType>(out->elemType());
-      if (elem_type != onnx::TensorProto::INT64 &&
-          elem_type != onnx::TensorProto::INT32) {
-        continue;
-      }
-      if (!out->has_sizes()) continue;
-      std::vector<int64_t> dims;
-      bool dims_known = true;
-      for (const auto& d : out->sizes()) {
-        if (!d.is_int) {
-          dims_known = false;
-          break;
-        }
-        dims.push_back(d.dim);
-      }
-      if (!dims_known) continue;
-      int64_t element_count = 1;
-      for (int64_t d : dims) element_count *= d;
-      if (element_count != static_cast<int64_t>(flat.size())) continue;
-      onnx::Tensor t;
-      t.elem_type() = elem_type;
-      t.sizes() = dims;
-      if (elem_type == onnx::TensorProto::INT64) {
-        for (int64_t v : flat) t.int64s().push_back(v);
-      } else {
-        for (int64_t v : flat) t.int32s().push_back(static_cast<int32_t>(v));
-      }
-      folded_values.emplace(output, std::move(t));
-    }
-    for (onnx::Node* node : node_ptrs) {
-      if (node->kind() != onnx::kReshape || node->inputs().size() < 2 ||
-          node->outputs().size() != 1) {
-        continue;
-      }
-      const std::string& reshape_out = node->outputs()[0]->uniqueName();
-      if (reshape_fixes.count(reshape_out)) continue;
-      auto sym_iter = sym_values.find(node->inputs()[1]->uniqueName());
-      if (sym_iter == sym_values.end()) continue;
-      const onnxsim::SymTensor& shape_value = sym_iter->second;
-      if (shape_value.scalar || shape_value.data.empty()) continue;
-      int unknown = 0;
-      bool usable = true;
-      std::vector<int64_t> shape;
-      shape.reserve(shape_value.data.size());
-      for (const auto& e : shape_value.data) {
-        if (e.is_symbolic()) {
-          ++unknown;
-          shape.push_back(-1);
-        } else {
-          const int64_t v = e.to_int();
-          if (v <= 0) {
-            usable = false;
-            break;
-          }
-          shape.push_back(v);
-        }
-      }
-      if (!usable || unknown != 1) continue;
-      onnx::Tensor t;
-      t.elem_type() = onnx::TensorProto::INT64;
-      t.sizes() = {static_cast<int64_t>(shape.size())};
-      for (int64_t v : shape) t.int64s().push_back(v);
-      reshape_fixes.emplace(reshape_out, ReshapeShapeFix{std::move(t)});
-    }
-  }
-
-  // Restore the graph's shape/type state to what it was before this pass's
-  // own data-propagation inference call -- see the snapshot comment above.
-  restore();
-
-  if (folded_values.empty() && reshape_fixes.empty()) {
-    return false;
-  }
-
-  // Apply the two rewrites directly on the Graph: a folded node is replaced
-  // in place by a Constant node holding its value (insertBefore + replaceAll
-  // UsesWith + destroy preserves topological position without a full
-  // node-list rebuild); a Reshape fix inserts a Constant just before the
-  // Reshape and repoints its shape input at it, leaving the original
-  // shape-producing subgraph dead for the optimizer's own dead-node
-  // elimination to remove.
-  static const onnx::Symbol kValueAttr("value");
-  for (onnx::Node* node : node_ptrs) {
-    if (node->outputs().size() != 1) continue;
-    onnx::Value* out = node->outputs()[0];
-    auto iter = folded_values.find(out->uniqueName());
-    if (iter != folded_values.end()) {
-      onnx::Node* constant = g.create(onnx::kConstant, 1);
-      constant->t_(kValueAttr, iter->second);
-      constant->outputs()[0]->setElemType(iter->second.elem_type());
-      constant->outputs()[0]->setSizes(std::vector<onnx::Dimension>(
-          iter->second.sizes().begin(), iter->second.sizes().end()));
-      constant->insertBefore(node);
-      node->replaceAllUsesWith(constant);
-      node->destroy();
-      continue;
-    }
-    auto fix_iter = reshape_fixes.find(out->uniqueName());
-    if (fix_iter != reshape_fixes.end() && node->kind() == onnx::kReshape) {
-      onnx::Node* shape_const = g.create(onnx::kConstant, 1);
-      shape_const->t_(kValueAttr, fix_iter->second.shape_tensor);
-      shape_const->outputs()[0]->setElemType(
-          fix_iter->second.shape_tensor.elem_type());
-      shape_const->outputs()[0]->setSizes(std::vector<onnx::Dimension>(
-          fix_iter->second.shape_tensor.sizes().begin(),
-          fix_iter->second.shape_tensor.sizes().end()));
-      shape_const->insertBefore(node);
-      node->replaceInput(1, shape_const->outputs()[0]);
-    }
-  }
-  return true;
-}
-
-// Whether every element of `tensor` is zero. Only the storage forms that can be
-// inspected locally are accepted: a tensor whose data lives in an external file
-// is reported as "not provably zero" rather than loaded.
-bool IsAllZeroTensor(const onnx::Tensor& tensor) {
-  if (tensor.data_location() == onnx::TensorProto_DataLocation_EXTERNAL) {
-    return false;
-  }
-  if (tensor.elem_type() == onnx::TensorProto::STRING ||
-      tensor.elem_type() == onnx::TensorProto::UNDEFINED) {
-    return false;
-  }
-  if (tensor.is_raw_data()) {
-    const std::string& raw = tensor.raw();
-    return std::all_of(raw.begin(), raw.end(), [](char c) { return c == 0; });
-  }
-  auto all_zero = [](const auto& field) {
-    return std::all_of(field.begin(), field.end(),
-                       [](auto value) { return value == 0; });
-  };
-  const size_t element_count =
-      tensor.floats().size() + tensor.int32s().size() + tensor.int64s().size() +
-      tensor.doubles().size() + tensor.uint64s().size();
-  if (element_count == 0) {
-    return false;
-  }
-  return all_zero(tensor.floats()) && all_zero(tensor.int32s()) &&
-         all_zero(tensor.int64s()) && all_zero(tensor.doubles()) &&
-         all_zero(tensor.uint64s());
-}
-
-// Whether `value` is provably an all-zero tensor, following the chain of ops
-// that produced it, walking onnx::Value/Node. Only shape-manipulating ops are
-// traversed: they move elements around without changing their values, so an
-// all-zero input implies an all-zero output. Ops whose output could be empty
-// (Slice, Split, Gather) stay sound too, since an empty tensor is vacuously
-// all zeros. ``value``'s producer being the graph's single kUndefined
-// placeholder node (see ir_pb_converter.cc) means "this optional input was
-// not provided".
-bool IsAllZeroGraphValue(
-    onnx::Value* value,
-    const std::unordered_map<std::string, const onnx::Tensor*>&
-        initializer_by_name,
-    std::unordered_map<const onnx::Value*, bool>& memo) {
-  if (value->node()->kind() == onnx::kUndefined) {
-    return false;
-  }
-  auto memo_iter = memo.find(value);
-  if (memo_iter != memo.end()) {
-    return memo_iter->second;
-  }
-  // Insert a pessimistic answer up front: it both memoizes the miss and
-  // breaks any cycle a malformed graph might contain.
-  memo.emplace(value, false);
-
-  auto init_iter = initializer_by_name.find(value->uniqueName());
-  if (init_iter != initializer_by_name.end()) {
-    const bool result = IsAllZeroTensor(*init_iter->second);
-    memo[value] = result;
-    return result;
-  }
-
-  onnx::Node* producer = value->node();
-  if (producer->has_domain() && !producer->domain().empty() &&
-      producer->domain() != "ai.onnx") {
-    return false;
-  }
-
-  static const onnx::Symbol kConstant("Constant");
-  static const onnx::Symbol kConstantOfShape("ConstantOfShape");
-  static const onnx::Symbol kCast("Cast");
-  static const onnx::Symbol kCastLike("CastLike");
-  static const onnx::Symbol kIdentity("Identity");
-  static const onnx::Symbol kReshape("Reshape");
-  static const onnx::Symbol kTranspose("Transpose");
-  static const onnx::Symbol kSqueeze("Squeeze");
-  static const onnx::Symbol kUnsqueeze("Unsqueeze");
-  static const onnx::Symbol kFlatten("Flatten");
-  static const onnx::Symbol kTile("Tile");
-  static const onnx::Symbol kExpand("Expand");
-  static const onnx::Symbol kSlice("Slice");
-  static const onnx::Symbol kSplit("Split");
-  static const onnx::Symbol kGather("Gather");
-  static const onnx::Symbol kGatherElements("GatherElements");
-  static const onnx::Symbol kGatherND("GatherND");
-  static const onnx::Symbol kConcat("Concat");
-  static const onnx::Symbol kValue("value");
-  static const onnx::Symbol kValueFloat("value_float");
-  static const onnx::Symbol kValueInt("value_int");
-  static const onnx::Symbol kValueFloats("value_floats");
-  static const onnx::Symbol kValueInts("value_ints");
-  static const onnx::Symbol kTo("to");
-
-  bool result = false;
-  const onnx::Symbol kind = producer->kind();
-  if (kind == kConstant) {
-    if (producer->hasAttribute(kValue)) {
-      result = IsAllZeroTensor(producer->t(kValue));
-    } else if (producer->hasAttribute(kValueFloat)) {
-      result = producer->f(kValueFloat) == 0;
-    } else if (producer->hasAttribute(kValueInt)) {
-      result = producer->i(kValueInt) == 0;
-    } else if (producer->hasAttribute(kValueFloats)) {
-      const auto& floats = producer->fs(kValueFloats);
-      result = !floats.empty() && std::all_of(floats.begin(), floats.end(),
-                                              [](double f) { return f == 0; });
-    } else if (producer->hasAttribute(kValueInts)) {
-      const auto& ints = producer->is(kValueInts);
-      result = !ints.empty() && std::all_of(ints.begin(), ints.end(),
-                                            [](int64_t i) { return i == 0; });
-    }
-  } else if (kind == kConstantOfShape) {
-    result =
-        !producer->hasAttribute(kValue) || IsAllZeroTensor(producer->t(kValue));
-  } else if (kind == kCast || kind == kCastLike) {
-    const bool to_string = producer->hasAttribute(kTo) &&
-                           producer->i(kTo) == onnx::TensorProto::STRING;
-    result = !to_string && IsAllZeroGraphValue(producer->inputs()[0],
-                                               initializer_by_name, memo);
-  } else if (kind == kIdentity || kind == kReshape || kind == kTranspose ||
-             kind == kSqueeze || kind == kUnsqueeze || kind == kFlatten ||
-             kind == kTile || kind == kExpand || kind == kSlice ||
-             kind == kSplit || kind == kGather || kind == kGatherElements ||
-             kind == kGatherND) {
-    result =
-        IsAllZeroGraphValue(producer->inputs()[0], initializer_by_name, memo);
-  } else if (kind == kConcat) {
-    const auto& inputs = producer->inputs();
-    result = !inputs.empty() &&
-             std::all_of(inputs.begin(), inputs.end(), [&](onnx::Value* input) {
-               return IsAllZeroGraphValue(input, initializer_by_name, memo);
-             });
-  }
-
-  memo[value] = result;
-  return result;
-}
-
-// The graph's single placeholder Value standing in for "this optional input
-// was not provided" (see ir_pb_converter.cc) -- every kUndefined-producer
-// input aliases this same Value. Only scanned when actually needed (an
-// all-zero RNN initial state was found), so the common case of a graph with
-// no LSTM/RNN/GRU nodes pays nothing for it.
-onnx::Value* FindUndefinedGraphValue(onnx::Graph& graph) {
-  for (onnx::Node* node : graph.nodes()) {
-    if (node->kind() == onnx::kUndefined) {
-      return node->outputs()[0];
-    }
-  }
-  return nullptr;
-}
-
-// Unset the recurrent initial states of RNN/GRU/LSTM nodes that are provably
-// all zeros (issue #314).
-//
-// paddle2onnx (like several other converters) materializes the zero initial
-// hidden/cell state of an LSTM as a *batch-dependent* subgraph, because the
-// state's shape is [num_directions, batch_size, hidden_size]:
-//   Shape(x) -> Slice -> Concat([batch,1,1]) -> Tile(zeros) -> Transpose
-//            -> Slice -> LSTM(initial_h, initial_c)
-// When the model has a dynamic batch dimension none of that can be constant
-// folded, so the simplified model keeps a Shape/Slice/Concat/Tile chain that
-// downstream converters (onnx2ncnn in the issue) reject outright.
-//
-// The ONNX spec says initial_h/initial_c default to zero when omitted, so an
-// input that is provably all zeros can simply be unset. The subgraph feeding it
-// then becomes dead and is removed by the ordinary dead-end elimination pass.
-// Only the initial states are unset; the equally zero-defaulting B and P inputs
-// are left alone because they are plain initializers, so dropping them removes
-// no operator while risking a needless behaviour change in consumers that read
-// them.
-void EliminateZeroRnnInitialState(onnx::Graph& graph) {
-  std::unordered_map<std::string, const onnx::Tensor*> initializer_by_name;
-  const auto& initializers = graph.initializers();
-  const auto& initializer_names = graph.initializer_names();
-  initializer_by_name.reserve(initializers.size());
-  for (size_t i = 0; i < initializers.size(); ++i) {
-    initializer_by_name[initializer_names[i]] = initializers[i].get();
-  }
-  std::unordered_map<const onnx::Value*, bool> memo;
-  onnx::Value* undefined = nullptr;
-
-  static const onnx::Symbol kLSTM("LSTM");
-  static const onnx::Symbol kRNN("RNN");
-  static const onnx::Symbol kGRU("GRU");
-
-  for (onnx::Node* node : graph.nodes()) {
-    // Recurse first, so recurrent ops inside If/Loop/Scan bodies are handled
-    // too.
-    for (onnx::Symbol attr : node->attributeNames()) {
-      if (node->kindOf(attr) == onnx::AttributeKind::g) {
-        EliminateZeroRnnInitialState(*node->g(attr));
-      } else if (node->kindOf(attr) == onnx::AttributeKind::gs) {
-        for (const auto& subgraph : node->gs(attr)) {
-          EliminateZeroRnnInitialState(*subgraph);
-        }
-      }
-    }
-
-    if (node->has_domain() && !node->domain().empty() &&
-        node->domain() != "ai.onnx") {
-      continue;
-    }
-    const onnx::Symbol kind = node->kind();
-    int last_state_index;
-    if (kind == kLSTM) {
-      last_state_index = 6;
-    } else if (kind == kRNN || kind == kGRU) {
-      last_state_index = 5;
-    } else {
-      continue;
-    }
-
-    const auto& inputs = node->inputs();
-    for (int i = 5;
-         i <= last_state_index && i < static_cast<int>(inputs.size()); i++) {
-      if (IsAllZeroGraphValue(inputs[i], initializer_by_name, memo)) {
-        if (undefined == nullptr) {
-          undefined = FindUndefinedGraphValue(graph);
-        }
-        node->replaceInput(i, undefined);
-      }
-    }
-    // Trailing empty inputs carry no information; drop them so the node ends
-    // up in the same shape a converter would have emitted without the state.
-    while (!node->inputs().empty() &&
-           node->inputs().back()->node()->kind() == onnx::kUndefined) {
-      node->removeInput(node->inputs().size() - 1);
-    }
-  }
-}
-
-// Estimate the number of bytes the outputs of `node` occupy once materialized,
-// using shapes gathered by shape inference (`vi_map` maps a tensor name to its
-// value_info). Outputs whose dtype or shape is not fully known contribute
-// nothing; the caller falls back to a node-count budget to stay bounded when no
-// size information is available.
-size_t EstimateOutputBytes(
-    const std::unordered_map<std::string, const onnx::ValueInfoProto*>& vi_map,
-    const onnx::NodeProto& node) {
-  size_t total = 0;
-  for (const auto& output : node.output()) {
-    auto iter = vi_map.find(output);
-    if (iter == vi_map.end() || !iter->second->type().has_tensor_type()) {
-      continue;
-    }
-    const auto& tensor_type = iter->second->type().tensor_type();
-    if (tensor_type.elem_type() == onnx::TensorProto::UNDEFINED) {
-      continue;
-    }
-    if (!tensor_type.has_shape()) {
-      continue;
-    }
-    size_t size;
-    try {
-      size = size_of_dtype(
-          static_cast<onnx::TensorProto::DataType>(tensor_type.elem_type()));
-    } catch (const std::exception&) {
-      // Unknown dtype: treat as unsized and rely on the node-count budget.
-      continue;
-    }
-    bool known = true;
-    for (const auto& dim : tensor_type.shape().dim()) {
-      if (!dim.has_dim_value()) {
-        known = false;
-        break;
-      }
-      size *= dim.dim_value();
-    }
-    if (known) {
-      total += size;
-    }
-  }
-  return total;
-}
-
-// Fold the const nodes in `const_nodes[begin, end)` as a single batch,
-// appending the resulting initializers to `model` and recording the folded
-// output names in `folded_outputs`. On failure the batch is bisected and each
-// half retried, so a single un-runnable node does not stop the rest of the
-// group from folding; the lower half is folded first and adds its initializers,
-// so the upper half can still read any values it depends on. A batch of one
-// that fails is skipped with a warning, matching the original per-node
-// behaviour.
-void FoldGroup(const ModelExecutor& executor, onnx::ModelProto& model,
-               const std::vector<onnx::NodeProto>& const_nodes, size_t begin,
-               size_t end,
-               const std::unordered_map<std::string, const onnx::NodeProto*>&
-                   deferred_producers,
-               std::set<std::string>& folded_outputs) {
-  if (begin >= end) {
-    return;
-  }
-  std::vector<const onnx::NodeProto*> ops;
-  ops.reserve(end - begin);
-  for (size_t k = begin; k < end; k++) {
-    ops.push_back(&const_nodes[k]);
-  }
-  try {
-    RunOpsAndAddInitializers(executor, model, ops, deferred_producers);
-    for (size_t k = begin; k < end; k++) {
-      for (const auto& output : const_nodes[k].output()) {
-        folded_outputs.insert(output);
-      }
-    }
-  } catch (const std::exception& e) {
-    if (end - begin == 1) {
-      const auto& x = const_nodes[begin];
-      std::cerr << "WARNING: failed to run \"" << x.op_type()
-                << "\" op (name is \"" << x.name() << "\"), skip... "
-                << e.what() << std::endl;
-      return;
-    }
-    const size_t mid = begin + (end - begin) / 2;
-    FoldGroup(executor, model, const_nodes, begin, mid, deferred_producers,
-              folded_outputs);
-    FoldGroup(executor, model, const_nodes, mid, end, deferred_producers,
-              folded_outputs);
-  }
-}
-
-// Takes `model` by value rather than `const&`: both call sites pass an
-// rvalue (std::move of a local they immediately overwrite with the return
-// value), so this is a cheap move-construction -- a pointer/buffer swap --
-// not a deep copy of the model's initializer bytes. `model` is then this
-// function's own uniquely-owned working copy, mutated in place throughout.
-onnx::ModelProto _FoldConstant(const ModelExecutor& executor,
-                               onnx::ModelProto model) {
-  ConstantNodePartition partition;
-  {
-    onnxsim::ProfiledScope analysis_scope("GetConstantNodes");
-    partition = GetConstantNodes(model);
-  }
-  const auto& const_nodes = partition.const_nodes;
-  // Map each deferred node's output to the producing node so RunOps can
-  // inline it into the sub-model executed when folding a downstream consumer.
-  // The pointers stay valid for the loop below: folding only appends
-  // initializers to the graph and never touches its node list.
-  std::unordered_map<std::string, const onnx::NodeProto*> deferred_producers;
-  if (!partition.deferred_outputs.empty()) {
-    for (const auto& node : model.graph().node()) {
-      for (const auto& output : node.output()) {
-        if (partition.deferred_outputs.count(output) > 0) {
-          deferred_producers.emplace(output, &node);
-        }
-      }
-    }
-  }
-  // Look up each tensor's inferred shape so batches can be capped by the
-  // bytes they would materialize (see below). Pointers reference `model`,
-  // which is not mutated (only appended to) while the map is in use.
-  std::unordered_map<std::string, const onnx::ValueInfoProto*> vi_map;
-  for (const auto& vi : model.graph().value_info()) {
-    vi_map[vi.name()] = &vi;
-  }
-  // Fold the const nodes in batches: one Session per batch instead of one per
-  // node. `const_nodes` is topologically sorted, so a batch is a contiguous
-  // slice and a later batch reads any earlier batch's outputs as ordinary
-  // initializers. Two budgets bound ORT's peak memory: a batch is closed once
-  // its outputs would exceed kBatchByteBudget or it reaches kBatchMaxNodes.
-  // Nodes that consume a deferred (large-tensor) output are folded on their
-  // own so the large intermediate is materialized transiently for just that
-  // node, exactly as in the per-node path.
-  constexpr size_t kBatchByteBudget = size_t(256) << 20;  // 256 MiB
-  constexpr size_t kBatchMaxNodes = 1024;
-  auto consumes_deferred = [&](const onnx::NodeProto& node) {
-    if (partition.deferred_outputs.empty()) {
-      return false;
-    }
-    for (const auto& input : node.input()) {
-      if (partition.deferred_outputs.count(input) > 0) {
-        return true;
-      }
-    }
-    return false;
-  };
-  // Outputs of const nodes that were successfully folded into initializers.
-  std::set<std::string> folded_outputs;
-  const size_t num_const_nodes = const_nodes.size();
-  for (size_t i = 0; i < num_const_nodes;) {
-    if (consumes_deferred(const_nodes[i])) {
-      FoldGroup(executor, model, const_nodes, i, i + 1, deferred_producers,
-                folded_outputs);
-      i++;
-      continue;
-    }
-    size_t j = i;
-    size_t bytes = 0;
-    while (j < num_const_nodes && j - i < kBatchMaxNodes &&
-           !consumes_deferred(const_nodes[j])) {
-      const size_t node_bytes = EstimateOutputBytes(vi_map, const_nodes[j]);
-      if (j > i && bytes + node_bytes > kBatchByteBudget) {
-        break;
-      }
-      bytes += node_bytes;
-      j++;
-    }
-    FoldGroup(executor, model, const_nodes, i, j, deferred_producers,
-              folded_outputs);
-    i = j;
-  }
-  // Rebuild the node list in its original topological order, dropping only
-  // the const nodes that were successfully folded into initializers. A const
-  // node that failed to fold must keep its original position: appending it to
-  // the end can place it after a non-const consumer (e.g. a Loop reading a
-  // SequenceEmpty output), which breaks topological sorting and makes the
-  // resulting model fail onnx's checker (issues #238, #335, #352).
-  {
-    onnxsim::ProfiledScope rebuild_scope("RebuildNodeList");
-    google::protobuf::RepeatedPtrField<onnx::NodeProto> original_nodes;
-    original_nodes.Swap(model.mutable_graph()->mutable_node());
-    for (auto& node : original_nodes) {
-      const bool folded =
-          node.output_size() > 0 && folded_outputs.count(node.output(0)) > 0;
-      if (!folded) {
-        *model.mutable_graph()->add_node() = std::move(node);
-      }
-    }
-  }
-  // Drop initializers left dangling by folding so the intermediate model does
-  // not balloon in size (issue #174).
-  onnxsim::ProfiledScope eliminate_scope("EliminateUnusedInitializerScope");
-  return EliminateUnusedInitializer(std::move(model));
-}
-
-// --- Graph-native counterpart of GetConstantNodes/FoldGroup/_FoldConstant ---
-//
-// Ports the ORT-execution-based constant folder above to walk onnx::Graph
-// directly, reached without a ModelProto <-> Graph round trip. Each fold
-// batch still serializes to a small, throwaway sub-ModelProto right at the
-// executor boundary (ORT has no Graph-native session API) -- exactly what
-// RunOps already did, just built from Graph Node/Value/Tensor objects
-// (reusing addAttribute/encodeTensor from ir_pb_converter_internal.h,
-// graph_shape_inference.cc's own pattern) instead of NodeProto copies.
-// Like the ModelProto path, _FoldConstantOnGraph runs onnx-optimizer's
-// eliminate_unused_initializer pass directly (not just via
-// config.optimizer_passes) at the end of every call: that pass list is
-// empty when the caller asked to skip optimization entirely
-// (perform_optimization=False), so nothing else is guaranteed to sweep up
-// an initializer folding leaves dangling in that mode.
-
-struct ConstantNodePartitionGraph {
-  // Nodes whose output is materialized into an initializer by folding, in
-  // topological order.
-  std::vector<onnx::Node*> const_nodes;
-  // Unique names of "deferred" nodes' outputs -- see ConstantNodePartition's
-  // own doc comment; same semantics, ported verbatim.
-  std::unordered_set<std::string> deferred_outputs;
-};
-
-bool HasSubgraphAttr(onnx::Node* node) {
-  for (onnx::Symbol name : node->attributeNames()) {
-    const auto kind = node->kindOf(name);
-    if (kind == onnx::AttributeKind::g || kind == onnx::AttributeKind::gs) {
-      return true;
-    }
-  }
-  return false;
-}
-
-// Graph-native counterpart of ProduceLargeTensor: reads the node's own
-// output Value directly instead of looking it up in a separate value_info
-// map (a Value already carries its own current shape/type in the IR).
-bool ProduceLargeTensorOnGraph(onnx::Node* node, size_t threshold) {
-  static const std::set<std::string> large_tensor_ops{"Tile", "ConstantOfShape",
-                                                      "Expand"};
-  if (large_tensor_ops.find(node->kind().toString()) ==
-      large_tensor_ops.end()) {
-    return false;
-  }
-  if (node->outputs().empty()) {
-    return true;
-  }
-  onnx::Value* out = node->outputs()[0];
-  if (out->elemType() == onnx::TensorProto::UNDEFINED || !out->has_sizes()) {
-    return true;
-  }
-  size_t size;
-  try {
-    size = size_of_dtype(
-        static_cast<onnx::TensorProto::DataType>(out->elemType()));
-  } catch (const std::exception&) {
-    return true;
-  }
-  for (const auto& d : out->sizes()) {
-    if (!d.is_int) {
-      return true;
-    }
-    size *= static_cast<size_t>(d.dim);
-  }
-  return size > threshold;
-}
-
-// Graph-native counterpart of EstimateOutputBytes.
-size_t EstimateOutputBytesOnGraph(onnx::Node* node) {
-  size_t total = 0;
-  for (onnx::Value* out : node->outputs()) {
-    if (out->elemType() == onnx::TensorProto::UNDEFINED || !out->has_sizes()) {
-      continue;
-    }
-    size_t size;
-    try {
-      size = size_of_dtype(
-          static_cast<onnx::TensorProto::DataType>(out->elemType()));
-    } catch (const std::exception&) {
-      continue;
-    }
-    bool known = true;
-    for (const auto& d : out->sizes()) {
-      if (!d.is_int) {
-        known = false;
-        break;
-      }
-      size *= static_cast<size_t>(d.dim);
-    }
-    if (known) {
-      total += size;
-    }
-  }
-  return total;
-}
-
-// Graph-native counterpart of GetConstantNodes.
-ConstantNodePartitionGraph GetConstantNodesOnGraph(
-    onnx::Graph& g, const std::vector<onnx::Node*>& node_ptrs) {
-  std::unordered_set<std::string> const_names{""};
-  ConstantNodePartitionGraph partition;
-  if (config.initializers_as_constants) {
-    for (const auto& name : g.initializer_names()) {
-      const_names.insert(name);
-    }
-  }
-  std::unordered_map<std::string, int> domain_to_version;
-  for (const onnx::OpSetID& opset : g.opset_versions_mutable()) {
-    const std::string& domain =
-        opset.domain() == "ai.onnx" ? "" : opset.domain();
-    domain_to_version[domain] = static_cast<int>(opset.version());
-  }
-  auto opset_version_of = [&domain_to_version](const std::string& domain) {
-    const std::string& key = domain == "ai.onnx" ? "" : domain;
-    auto iter = domain_to_version.find(key);
-    return iter == domain_to_version.end() ? 0 : iter->second;
-  };
-  for (onnx::Node* node : node_ptrs) {
-    if (node->kind() == onnx::kUndefined || node->kind() == onnx::kCaptured) {
-      continue;
-    }
-    const std::string domain =
-        node->has_domain() ? node->domain() : std::string();
-    const std::string op_type = node->kind().toString();
-    const bool foldable =
-        IsOfficialOp(domain, op_type) &&
-        IsDeterministic(domain, op_type, opset_version_of(domain)) &&
-        !IsQDQ(domain, op_type) && !HasSubgraphAttr(node) &&
-        std::all_of(
-            node->inputs().begin(), node->inputs().end(),
-            [&const_names](onnx::Value* v) {
-              const std::string name =
-                  v->node()->kind() == onnx::kUndefined ? "" : v->uniqueName();
-              return const_names.count(name) > 0;
-            });
-    if (!foldable) {
-      continue;
-    }
-    if (!ProduceLargeTensorOnGraph(node, config.tensor_size_threshold)) {
-      for (onnx::Value* out : node->outputs()) {
-        const_names.insert(out->uniqueName());
-      }
-      partition.const_nodes.push_back(node);
-      continue;
-    }
-    if (op_type == "ConstantOfShape" || op_type == "Expand") {
-      for (onnx::Value* out : node->outputs()) {
-        const_names.insert(out->uniqueName());
-        partition.deferred_outputs.insert(out->uniqueName());
-      }
-    }
-  }
-  return partition;
-}
-
-// Graph-native counterpart of RunOps: builds the throwaway sub-model from
-// Graph Node/Value/Tensor objects. Each constant feed is encoded into its
-// own owned TensorProto (encodeTensor copies its bytes once) and handed to
-// the DLPack bridge via FromTensorProtoOwning, which takes ownership of the
-// moved proto -- safe regardless of the source onnx::Tensor's lifetime,
-// unlike borrowing a pointer into a temporary.
-struct RunOpsOnGraphResult {
-  std::vector<onnx::TensorProto> tensors;
-  // Parallel to `tensors`: the original graph Value each entry replaces
-  // (the node output RunOps computed a constant for).
-  std::vector<onnx::Value*> values;
-};
-
-RunOpsOnGraphResult RunOpsOnGraph(
-    const ModelExecutor& executor, onnx::Graph& g,
-    const std::vector<onnx::Node*>& ops,
-    const std::unordered_map<std::string, onnx::Node*>& deferred_producers,
-    int64_t ir_version) {
-  std::vector<std::string> input_names;
-  std::vector<const onnx::Tensor*> input_tensors;
-  std::set<std::string> seen_inputs;
-
-  google::protobuf::Arena arena;
-  onnx::ModelProto& op_model =
-      *google::protobuf::Arena::Create<onnx::ModelProto>(&arena);
-  // Inherited from the model being simplified (not onnx::Version::IR_VERSION,
-  // this library's own compiled-in constant): onnxsim's vendored onnx tracks
-  // upstream's IR_VERSION ahead of what any released onnxruntime actually
-  // supports, and this throwaway sub-model still has to load in a real
-  // onnxruntime session for its one Run() call -- the original model's own
-  // ir_version is what the caller's onnxruntime install already proved it
-  // could handle.
-  op_model.set_ir_version(ir_version);
-  for (const onnx::OpSetID& opset : g.opset_versions_mutable()) {
-    auto* x = op_model.add_opset_import();
-    x->set_domain(opset.domain());
-    x->set_version(opset.version());
-  }
-
-  std::set<std::string> internal_outputs;
-  for (onnx::Node* op : ops) {
-    for (onnx::Value* out : op->outputs()) {
-      internal_outputs.insert(out->uniqueName());
-    }
-  }
-
-  std::set<onnx::Node*> included;
-  std::function<void(onnx::Node*)> include_node = [&](onnx::Node* node) {
-    if (!included.insert(node).second) {
-      return;
-    }
-    for (onnx::Value* input : node->inputs()) {
-      if (input->node()->kind() == onnx::kUndefined) {
-        continue;  // unset optional input
-      }
-      const std::string& name = input->uniqueName();
-      if (internal_outputs.count(name) > 0) {
-        continue;
-      }
-      auto deferred_iter = deferred_producers.find(name);
-      if (deferred_iter != deferred_producers.end()) {
-        include_node(deferred_iter->second);
-        continue;
-      }
-      if (!seen_inputs.insert(name).second) {
-        continue;
-      }
-      const onnx::Tensor* init = g.getInitializer(name);
-      if (init == nullptr) {
-        // Mirrors FindInitializerByName's own throwing behavior: `name`
-        // was classified constant during GetConstantNodesOnGraph's static
-        // partition, but its producer never actually materialized an
-        // initializer for it (e.g. an earlier batch's fold for it failed
-        // and was skipped, see FoldGroupOnGraph's catch clause). Throwing
-        // here lets the same catch-and-bisect/skip machinery handle this
-        // node too, instead of crashing.
-        throw std::invalid_argument("no initializer " + name);
-      }
-      if (init->sizes().size() == 1 && init->sizes()[0] == 0) {
-        onnx::TensorProto* p = op_model.mutable_graph()->add_initializer();
-        p->set_name(name);
-        onnx::encodeTensor(*p, *init);
-        continue;
-      }
-      input_names.push_back(name);
-      input_tensors.push_back(init);
-    }
-    onnx::NodeProto* np = op_model.mutable_graph()->add_node();
-    np->set_op_type(node->kind().toString());
-    if (node->has_domain()) {
-      np->set_domain(node->domain());
-    }
-    for (onnx::Value* input : node->inputs()) {
-      np->add_input(
-          input->node()->kind() == onnx::kUndefined ? "" : input->uniqueName());
-    }
-    for (onnx::Value* output : node->outputs()) {
-      np->add_output(output->uniqueName());
-    }
-    for (onnx::Symbol attr_name : node->attributeNames()) {
-      onnx::addAttribute(*np, *node, attr_name, /*consume_tensor_data=*/false);
-    }
-  };
-  for (onnx::Node* op : ops) {
-    include_node(op);
-  }
-
-  for (size_t i = 0; i < input_names.size(); i++) {
-    onnx::ValueInfoProto* vi = op_model.mutable_graph()->add_input();
-    vi->set_name(input_names[i]);
-    auto* tensor_type = vi->mutable_type()->mutable_tensor_type();
-    tensor_type->set_elem_type(input_tensors[i]->elem_type());
-    for (int64_t d : input_tensors[i]->sizes()) {
-      tensor_type->mutable_shape()->add_dim()->set_dim_value(d);
-    }
-  }
-  std::vector<std::string> output_names;
-  std::vector<onnx::Value*> output_values;
-  for (onnx::Node* op : ops) {
-    for (onnx::Value* out : op->outputs()) {
-      op_model.mutable_graph()->add_output()->set_name(out->uniqueName());
-      output_names.push_back(out->uniqueName());
-      output_values.push_back(out);
-    }
-  }
-
-  std::vector<DLManagedTensorPtr> input_dls;
-  std::vector<const DLManagedTensor*> input_ptrs;
-  input_dls.reserve(input_tensors.size());
-  for (const auto* t : input_tensors) {
-    onnx::TensorProto tp;
-    onnx::encodeTensor(tp, *t);
-    input_dls.emplace_back(
-        onnxsim::dlpack::FromTensorProtoOwning(std::move(tp)));
-  }
-  input_ptrs.reserve(input_dls.size());
-  for (const auto& p : input_dls) input_ptrs.push_back(p.get());
-
-  std::vector<DLManagedTensorPtr> output_dls;
-  {
-    // Profiled under the same "OrtSession" span name as the ModelProto-based
-    // RunOps (see its own comment): the one spot common to every executor
-    // binding, so a trace can find the actual session-run time regardless of
-    // which folding path produced it.
-    onnxsim::ProfiledScope session_scope("OrtSession");
-    output_dls = executor.Run(op_model, input_ptrs);
-  }
-  RunOpsOnGraphResult result;
-  result.tensors.reserve(output_dls.size());
-  for (size_t i = 0; i < output_dls.size(); i++) {
-    result.tensors.push_back(onnxsim::dlpack::ToTensorProto(
-        output_dls[i]->dl_tensor,
-        i < output_names.size() ? output_names[i] : std::string()));
-  }
-  // executor.Run() returns one tensor per graph output, in the same order
-  // output_values (built alongside output_names above) lists them in.
-  result.values = output_values;
-  result.values.resize(result.tensors.size());
-  return result;
-}
-
-// Graph-native counterpart of FoldGroup: splices each successfully-folded
-// output directly into `g` as a new initializer (Graph::
-// addInitializerAndCreateValue), rewires the folded node's uses onto it,
-// and removes the folded node -- no NodeProto rebuild needed.
-void FoldGroupOnGraph(
-    const ModelExecutor& executor, onnx::Graph& g,
-    const std::vector<onnx::Node*>& const_nodes, size_t begin, size_t end,
-    const std::unordered_map<std::string, onnx::Node*>& deferred_producers,
-    size_t& num_folded, int64_t ir_version) {
-  if (begin >= end) {
-    return;
-  }
-  std::vector<onnx::Node*> ops(const_nodes.begin() + begin,
-                               const_nodes.begin() + end);
-  try {
-    RunOpsOnGraphResult result =
-        RunOpsOnGraph(executor, g, ops, deferred_producers, ir_version);
-    // Every op in this batch folded successfully (RunOpsOnGraph throws,
-    // rather than partially populating its result, on any failure) -- decode
-    // each returned TensorProto (ToTensorProto always emits raw_data, see
-    // dlpack_bridge.h) into a Tensor, add it as a new initializer, and
-    // rewire the Value it replaces onto it. A multi-output node's outputs
-    // are independent Values here, each replaced on its own; the owning
-    // node is only destroyed (never touched again after) once every one of
-    // its outputs has been replaced.
-    std::unordered_map<onnx::Node*, size_t> remaining_outputs;
-    for (onnx::Node* node : ops) {
-      remaining_outputs[node] = node->outputs().size();
-    }
-    for (size_t i = 0; i < result.tensors.size(); i++) {
-      const onnx::TensorProto& tp = result.tensors[i];
-      onnx::Value* old_value = result.values[i];
-      onnx::Node* owner = old_value->node();
-      onnx::Tensor t;
-      t.setName(tp.name());
-      t.elem_type() = tp.data_type();
-      for (int64_t d : tp.dims()) t.sizes().push_back(d);
-      if (tp.has_raw_data()) {
-        t.set_raw_data(tp.raw_data());
-      }
-      onnx::Value* new_value = g.addInitializerAndCreateValue(t);
-      old_value->replaceAllUsesWith(new_value);
-      if (--remaining_outputs[owner] == 0) {
-        owner->destroy();
-      }
-    }
-    num_folded += end - begin;
-  } catch (const std::exception& e) {
-    if (end - begin == 1) {
-      onnx::Node* node = const_nodes[begin];
-      std::cerr << "WARNING: failed to run \"" << node->kind().toString()
-                << "\" op (name is \"" << node->name() << "\"), skip... "
-                << e.what() << std::endl;
-      return;
-    }
-    const size_t mid = begin + (end - begin) / 2;
-    FoldGroupOnGraph(executor, g, const_nodes, begin, mid, deferred_producers,
-                     num_folded, ir_version);
-    FoldGroupOnGraph(executor, g, const_nodes, mid, end, deferred_producers,
-                     num_folded, ir_version);
-  }
-}
-
-// Graph-native counterpart of _FoldConstant. Returns whether anything
-// folded, mirroring OptimizeGraphChanged's own bool-returning convention so
-// the outer fixed point can detect convergence without a fingerprint
-// comparison. Always sweeps unused initializers at the end (see this
-// section's own top-of-file comment for why).
-bool _FoldConstantOnGraph(const ModelExecutor& executor, onnx::Graph& g,
-                          int64_t ir_version) {
-  std::vector<onnx::Node*> node_ptrs(g.nodes().begin(), g.nodes().end());
-  ConstantNodePartitionGraph partition;
-  {
-    onnxsim::ProfiledScope analysis_scope("GetConstantNodes");
-    partition = GetConstantNodesOnGraph(g, node_ptrs);
-  }
-  const auto& const_nodes = partition.const_nodes;
-  if (const_nodes.empty()) {
-    // Nothing to fold this call, but a dangling initializer could already be
-    // sitting in `g` from a previous round or the input model itself -- see
-    // this function's own doc comment for why this sweep cannot be skipped.
-    onnx::optimization::EliminateUnusedInitializer()
-        .eliminate_unused_initializer(g);
-    return false;
-  }
-
-  std::unordered_map<std::string, onnx::Node*> deferred_producers;
-  if (!partition.deferred_outputs.empty()) {
-    for (onnx::Node* node : node_ptrs) {
-      for (onnx::Value* out : node->outputs()) {
-        if (partition.deferred_outputs.count(out->uniqueName()) > 0) {
-          deferred_producers.emplace(out->uniqueName(), node);
-        }
-      }
-    }
-  }
-
-  constexpr size_t kBatchByteBudget = size_t(256) << 20;  // 256 MiB
-  constexpr size_t kBatchMaxNodes = 1024;
-  auto consumes_deferred = [&](onnx::Node* node) {
-    if (partition.deferred_outputs.empty()) {
-      return false;
-    }
-    for (onnx::Value* input : node->inputs()) {
-      if (partition.deferred_outputs.count(input->uniqueName()) > 0) {
-        return true;
-      }
-    }
-    return false;
-  };
-
-  size_t num_folded = 0;
-  const size_t num_const_nodes = const_nodes.size();
-  for (size_t i = 0; i < num_const_nodes;) {
-    if (consumes_deferred(const_nodes[i])) {
-      FoldGroupOnGraph(executor, g, const_nodes, i, i + 1, deferred_producers,
-                       num_folded, ir_version);
-      i++;
-      continue;
-    }
-    size_t j = i;
-    size_t bytes = 0;
-    while (j < num_const_nodes && j - i < kBatchMaxNodes &&
-           !consumes_deferred(const_nodes[j])) {
-      const size_t node_bytes = EstimateOutputBytesOnGraph(const_nodes[j]);
-      if (j > i && bytes + node_bytes > kBatchByteBudget) {
-        break;
-      }
-      bytes += node_bytes;
-      j++;
-    }
-    FoldGroupOnGraph(executor, g, const_nodes, i, j, deferred_producers,
-                     num_folded, ir_version);
-    i = j;
-  }
-  // Drop initializers left dangling by folding so the graph does not balloon
-  // in size (issue #174) -- see this function's own doc comment for why this
-  // cannot be left to config.optimizer_passes alone.
-  onnx::optimization::EliminateUnusedInitializer().eliminate_unused_initializer(
-      g);
-  return num_folded > 0;
-}
-
-// ``model`` is ``onnx::ModelProto&`` (not const): the call site below passes a
-// mutable lvalue that is about to be move-assigned over
-// (``model = Optimize(model)``), so its pre-call contents are dead once this
-// returns. That lets OptimizeFixed move each initializer's raw bytes through
-// the ModelProto<->Graph round trip instead of copying them -- the dominant
-// cost of this call for weight-heavy models (onnxsim issue #633) -- via the
-// moving ImportModelProto/ExportModelProto overloads added to onnxsim's own
-// onnx fork.
-onnx::ModelProto Optimize(onnx::ModelProto& model) {
-  // Make onnxsim's own optimizer passes available to onnxoptimizer's registry
-  // (idempotent) so config.optimizer_passes may name them.
-  onnxsim::RegisterCustomOptimizerPasses();
-  // Mirror the initializer treatment into the onnx optimizer so its
-  // value-baking passes (fuse_bn_into_conv, ...) respect it too. The setting is
-  // thread-local in the optimizer; restore it afterwards so we do not leak it.
-  const bool prev = onnx::optimization::InitializersAsConstants();
-  onnx::optimization::SetInitializersAsConstants(
-      config.initializers_as_constants);
-  auto result =
-      onnx::optimization::OptimizeFixed(model, config.optimizer_passes);
-  onnx::optimization::SetInitializersAsConstants(prev);
-  return result;
-}
 
 // A 128-bit fingerprint of a model, used by FixedPointFn to detect when an
 // iteration stopped changing the model without keeping a second full ModelProto
